Methods and Models for Claims Reserving

ChainLadder provides methods and models which are typically used in insurance claims reserving.

The package grew out of presentations given at the Stochastic Reserving Seminar at the Institute of Actuaries in 2007 and 2008 and followed by talks at CAS meetings in 2008 and 2010.

More information is available on the project web site https://github.com/mages/ChainLadder

For more financial packages see also CRAN Task View 'Emperical Finance' at https://CRAN.R-project.org/view=Finance.

Author

Maintainer: Markus Gesmann <markus.gesmann@gmail.com>

References

Thomas Mack. Distribution-free calculation of the standard error of chain ladder reserve estimates. Astin Bulletin. Vol. 23. No 2. 1993. pp.213:225

Thomas Mack. The standard error of chain ladder reserve estimates: Recursive calculation and inclusion of a tail factor. Astin Bulletin. Vol. 29. No 2. 1999. pp.361:366

Gerhard Quarg and Thomas Mack. Munich Chain Ladder. Blatter DGVFM 26. Munich. 2004.

England, PD and Verrall, RJ. Stochastic Claims Reserving in General Insurance (with discussion). British Actuarial Journal 8. III. 2002

B. Zehnwirth and G. Barnett. Best Estimates for Reserves. Proceedings of the CAS. Volume LXXXVII. Number 167.November 2000.

Clark, David R., "LDF Curve-Fitting and Stochastic Reserving: A Maximum Likelihood Approach," CAS Forum, Fall 2003.

Zhang Y. A general multivariate chain ladder model.Insurance: Mathematics and Economics, 46, pp. 588:599, 2010.

Zhang, Y. Likelihood-based and Bayesian Methods for Tweedie Compound Poisson Linear Mixed Models, Statistics and Computing, forthcoming.

Bardis, Majidi, Murphy. A Family of Chain-Ladder Factor Models for Selected Link Ratios. Variance. Pending. Variance 6:2, 2012, pp. 143-160.

Modelling the claims development result for solvency purposes. Michael Merz, Mario V. Wüthrich. Casualty Actuarial Society E-Forum, Fall 2008.

Claims Run-Off Uncertainty: The Full Picture. Michael Merz, Mario V. Wüthrich. Swiss Finance Institute Research Paper No. 14-69. 2014

Markus Gesmann. Claims Reserving and IBNR. Computational Actuarial Science with R. Chapman and Hall/CRC. 2014

Examples

if (FALSE) { # \dontrun{
  demo(ChainLadder)
  } # }