Example paid and incurred triangle data from CAS web site.

Paid and incurred triangle data from the United Services Automobile Association company for the private passenger auto liability/medical line of business.

Usage

data("USAApaid")

Format

A triangle with 10 accident years and 10 development years.

Details

The claims data comes from Schedule P - Analysis of Losses and Loss Expenses in the National Association of Insurance Commissioners (NAIC) database. CAS obtained permission from the NAIC to make this data available to all interested researchers on the CAS website. NAIC Schedule P contains information on claims for major personal and commercial lines for all property-casualty insurers that write business in US.

Source

Casualty Actuarial Society

References

CAS website: https://www.casact.org

Examples

data(USAApaid)