Methods for BootChainLadder objects

summary, print, mean, and quantile methods for BootChainLadder objects

Usage

# S3 method for BootChainLadder
summary(object, probs=c(0.75,0.95), ...)

# S3 method for BootChainLadder
print(x, probs=c(0.75,0.95), ...)

# S3 method for BootChainLadder
quantile(x, probs=c(0.75, 0.95), na.rm = FALSE,
              names = TRUE, type = 7,...)

# S3 method for BootChainLadder
mean(x, ...)

# S3 method for BootChainLadder
residuals(object, ...)

Arguments

x, object

output from BootChainLadder

probs

numeric vector of probabilities with values in [0,1], see quantile for more help

na.rm

logical; if true, any NA and NaN's are removed from 'x' before the quantiles are computed, see quantile for more help

names

logical; if true, the result has a names attribute. Set to FALSE for speedup with many 'probs', see quantile for more help

type

an integer between 1 and 9 selecting one of the nine quantile algorithms detailed below to be used, see quantile

...

further arguments passed to or from other methods

Details

print.BootChainLadder calls summary.BootChainLadder and prints a formatted version of the summary. residuals.BootChainLadder gives the residual triangle of the expected chain-ladder minus the actual triangle back.

Value

summary.BootChainLadder, mean.BootChainLadder, and

quantile.BootChainLadder, give a list with two elements back:

ByOrigin

data frame with summary/mean/quantile statistics by origin period

Totals

data frame with total summary/mean/quantile statistics for all origin period

Author

Markus Gesmann

See also

Examples

B <- BootChainLadder(RAA, R=999, process.distr="gamma")
B
#> BootChainLadder(Triangle = RAA, R = 999, process.distr = "gamma")
#> 
#>      Latest Mean Ultimate Mean IBNR IBNR.S.E IBNR 75% IBNR 95%
#> 1981 18,834        18,834         0        0        0        0
#> 1982 16,704        16,909       205      735      235    1,687
#> 1983 23,466        24,080       614    1,271    1,113    2,988
#> 1984 27,067        28,829     1,762    1,985    2,747    5,779
#> 1985 26,180        29,108     2,928    2,361    4,233    7,380
#> 1986 15,852        19,520     3,668    2,386    5,068    8,262
#> 1987 12,314        17,775     5,461    3,187    7,429   11,304
#> 1988 13,112        24,446    11,334    4,925   14,185   20,005
#> 1989  5,395        16,513    11,118    5,937   14,626   22,210
#> 1990  2,063        19,693    17,630   14,729   24,919   43,410
#> 
#>                  Totals
#> Latest:         160,987
#> Mean Ultimate:  215,706
#> Mean IBNR:       54,719
#> IBNR.S.E         19,498
#> Total IBNR 75%:  65,689
#> Total IBNR 95%:  89,108
summary(B)
#> $ByOrigin
#>      Latest Mean Ultimate  Mean IBNR    SD IBNR   IBNR 75%  IBNR 95%
#> 1981  18834      18834.00     0.0000     0.0000     0.0000     0.000
#> 1982  16704      16908.86   204.8574   735.4097   235.2108  1687.469
#> 1983  23466      24080.07   614.0671  1270.8300  1113.2651  2987.549
#> 1984  27067      28829.02  1762.0183  1984.8529  2747.1522  5779.282
#> 1985  26180      29108.10  2928.0974  2361.0936  4233.0899  7380.450
#> 1986  15852      19519.55  3667.5478  2385.7452  5068.3213  8261.823
#> 1987  12314      17774.74  5460.7379  3187.0584  7428.8237 11304.486
#> 1988  13112      24445.63 11333.6326  4924.8642 14184.5109 20005.043
#> 1989   5395      16513.47 11118.4716  5936.5527 14625.5499 22209.584
#> 1990   2063      19692.79 17629.7909 14728.6421 24918.6218 43409.758
#> 
#> $Totals
#>                    Totals
#> Latest:         160987.00
#> Mean Ultimate:  215706.22
#> Mean IBNR:       54719.22
#> SD IBNR:         19497.82
#> Total IBNR 75%:  65689.02
#> Total IBNR 95%:  89108.21
#> 
mean(B)
#> $ByOrigin
#>       Mean IBNR
#> 1981     0.0000
#> 1982   204.8574
#> 1983   614.0671
#> 1984  1762.0183
#> 1985  2928.0974
#> 1986  3667.5478
#> 1987  5460.7379
#> 1988 11333.6326
#> 1989 11118.4716
#> 1990 17629.7909
#> 
#> $Totals
#>               Total
#> Mean IBNR: 54719.22
#> 
quantile(B, c(0.75,0.95,0.99, 0.995))
#> $ByOrigin
#>        IBNR 75%  IBNR 95%  IBNR 99% IBNR 99.5%
#> 1981     0.0000     0.000     0.000      0.000
#> 1982   235.2108  1687.469  2806.677   3190.532
#> 1983  1113.2651  2987.549  5009.520   5698.584
#> 1984  2747.1522  5779.282  7788.197   8547.186
#> 1985  4233.0899  7380.450 10330.148  11763.865
#> 1986  5068.3213  8261.823 10135.783  10500.995
#> 1987  7428.8237 11304.486 14491.126  16120.450
#> 1988 14184.5109 20005.043 25104.749  29455.667
#> 1989 14625.5499 22209.584 27377.251  29537.209
#> 1990 24918.6218 43409.758 65819.189  69349.626
#> 
#> $Totals
#>                Totals
#> IBNR 75%:    65689.02
#> IBNR 95%:    89108.21
#> IBNR 99%:   111598.13
#> IBNR 99.5%: 116304.76
#>