Methods for BootChainLadder objects

summary, print, mean, and quantile methods for BootChainLadder objects

Usage

# S3 method for BootChainLadder
summary(object, probs=c(0.75,0.95), ...)

# S3 method for BootChainLadder
print(x, probs=c(0.75,0.95), ...)

# S3 method for BootChainLadder
quantile(x, probs=c(0.75, 0.95), na.rm = FALSE,
              names = TRUE, type = 7,...)

# S3 method for BootChainLadder
mean(x, ...)

# S3 method for BootChainLadder
residuals(object, ...)

Arguments

x, object

output from BootChainLadder

probs

numeric vector of probabilities with values in [0,1], see quantile for more help

na.rm

logical; if true, any NA and NaN's are removed from 'x' before the quantiles are computed, see quantile for more help

names

logical; if true, the result has a names attribute. Set to FALSE for speedup with many 'probs', see quantile for more help

type

an integer between 1 and 9 selecting one of the nine quantile algorithms detailed below to be used, see quantile

...

further arguments passed to or from other methods

Details

print.BootChainLadder calls summary.BootChainLadder and prints a formatted version of the summary. residuals.BootChainLadder gives the residual triangle of the expected chain-ladder minus the actual triangle back.

Value

summary.BootChainLadder, mean.BootChainLadder, and

quantile.BootChainLadder, give a list with two elements back:

ByOrigin

data frame with summary/mean/quantile statistics by origin period

Totals

data frame with total summary/mean/quantile statistics for all origin period

Author

Markus Gesmann

See also

Examples

B <- BootChainLadder(RAA, R=999, process.distr="gamma")
B
#> BootChainLadder(Triangle = RAA, R = 999, process.distr = "gamma")
#> 
#>      Latest Mean Ultimate Mean IBNR IBNR.S.E IBNR 75% IBNR 95%
#> 1981 18,834        18,834         0        0        0        0
#> 1982 16,704        16,869       165      678      202    1,344
#> 1983 23,466        24,130       664    1,369    1,136    3,335
#> 1984 27,067        28,856     1,789    2,002    2,681    5,441
#> 1985 26,180        28,980     2,800    2,303    4,155    7,209
#> 1986 15,852        19,604     3,752    2,443    5,073    8,225
#> 1987 12,314        17,721     5,407    2,970    6,993   10,827
#> 1988 13,112        24,316    11,204    4,984   14,316   19,433
#> 1989  5,395        16,228    10,833    6,042   14,512   22,565
#> 1990  2,063        20,250    18,187   14,114   25,691   45,564
#> 
#>                  Totals
#> Latest:         160,987
#> Mean Ultimate:  215,788
#> Mean IBNR:       54,801
#> IBNR.S.E         18,512
#> Total IBNR 75%:  64,805
#> Total IBNR 95%:  88,806
summary(B)
#> $ByOrigin
#>      Latest Mean Ultimate  Mean IBNR    SD IBNR   IBNR 75%  IBNR 95%
#> 1981  18834      18834.00     0.0000     0.0000     0.0000     0.000
#> 1982  16704      16868.56   164.5640   678.3205   201.5303  1343.707
#> 1983  23466      24130.31   664.3122  1368.9201  1136.4090  3335.408
#> 1984  27067      28855.87  1788.8721  2001.6180  2681.1935  5440.883
#> 1985  26180      28979.83  2799.8291  2303.4567  4154.5064  7209.384
#> 1986  15852      19604.23  3752.2331  2443.3810  5072.6958  8225.029
#> 1987  12314      17721.07  5407.0652  2970.3800  6992.8648 10827.417
#> 1988  13112      24316.06 11204.0551  4984.1377 14315.6444 19433.202
#> 1989   5395      16227.96 10832.9558  6042.4369 14511.6647 22565.103
#> 1990   2063      20249.91 18186.9136 14114.2905 25690.6433 45564.135
#> 
#> $Totals
#>                    Totals
#> Latest:         160987.00
#> Mean Ultimate:  215787.80
#> Mean IBNR:       54800.80
#> SD IBNR:         18512.47
#> Total IBNR 75%:  64805.25
#> Total IBNR 95%:  88805.67
#> 
mean(B)
#> $ByOrigin
#>       Mean IBNR
#> 1981     0.0000
#> 1982   164.5640
#> 1983   664.3122
#> 1984  1788.8721
#> 1985  2799.8291
#> 1986  3752.2331
#> 1987  5407.0652
#> 1988 11204.0551
#> 1989 10832.9558
#> 1990 18186.9136
#> 
#> $Totals
#>              Total
#> Mean IBNR: 54800.8
#> 
quantile(B, c(0.75,0.95,0.99, 0.995))
#> $ByOrigin
#>        IBNR 75%  IBNR 95%  IBNR 99% IBNR 99.5%
#> 1981     0.0000     0.000     0.000      0.000
#> 1982   201.5303  1343.707  2759.438   3019.325
#> 1983  1136.4090  3335.408  5494.323   6324.854
#> 1984  2681.1935  5440.883  8191.958   8857.511
#> 1985  4154.5064  7209.384  9500.330  10607.918
#> 1986  5072.6958  8225.029 10871.707  12872.259
#> 1987  6992.8648 10827.417 14510.762  14895.874
#> 1988 14315.6444 19433.202 25515.945  29081.560
#> 1989 14511.6647 22565.103 28453.699  29472.039
#> 1990 25690.6433 45564.135 60818.498  64566.249
#> 
#> $Totals
#>                Totals
#> IBNR 75%:    64805.25
#> IBNR 95%:    88805.67
#> IBNR 99%:   108533.21
#> IBNR 99.5%: 114795.43
#>