Methods for BootChainLadder objects

summary, print, mean, and quantile methods for BootChainLadder objects

Usage

# S3 method for BootChainLadder
summary(object, probs=c(0.75,0.95), ...)

# S3 method for BootChainLadder
print(x, probs=c(0.75,0.95), ...)

# S3 method for BootChainLadder
quantile(x, probs=c(0.75, 0.95), na.rm = FALSE,
              names = TRUE, type = 7,...)

# S3 method for BootChainLadder
mean(x, ...)

# S3 method for BootChainLadder
residuals(object, ...)

Arguments

x, object

output from BootChainLadder

probs

numeric vector of probabilities with values in [0,1], see quantile for more help

na.rm

logical; if true, any NA and NaN's are removed from 'x' before the quantiles are computed, see quantile for more help

names

logical; if true, the result has a names attribute. Set to FALSE for speedup with many 'probs', see quantile for more help

type

an integer between 1 and 9 selecting one of the nine quantile algorithms detailed below to be used, see quantile

...

further arguments passed to or from other methods

Details

print.BootChainLadder calls summary.BootChainLadder and prints a formatted version of the summary. residuals.BootChainLadder gives the residual triangle of the expected chain-ladder minus the actual triangle back.

Value

summary.BootChainLadder, mean.BootChainLadder, and quantile.BootChainLadder, give a list with two elements back:

ByOrigin

data frame with summary/mean/quantile statistics by origin period

Totals

data frame with total summary/mean/quantile statistics for all origin period

Author

Markus Gesmann

See also

Examples

B <- BootChainLadder(RAA, R=999, process.distr="gamma")
B
#> BootChainLadder(Triangle = RAA, R = 999, process.distr = "gamma")
#> 
#>      Latest Mean Ultimate Mean IBNR IBNR.S.E IBNR 75% IBNR 95%
#> 1981 18,834        18,834         0        0        0        0
#> 1982 16,704        16,860       156      691      188    1,416
#> 1983 23,466        24,124       658    1,336    1,037    3,333
#> 1984 27,067        28,761     1,694    1,851    2,508    5,441
#> 1985 26,180        29,016     2,836    2,330    4,209    7,128
#> 1986 15,852        19,647     3,795    2,580    5,068    8,810
#> 1987 12,314        17,650     5,336    3,165    6,907   11,093
#> 1988 13,112        24,208    11,096    5,014   13,914   20,546
#> 1989  5,395        16,231    10,836    6,030   14,421   22,058
#> 1990  2,063        19,557    17,494   13,885   24,849   43,075
#> 
#>                  Totals
#> Latest:         160,987
#> Mean Ultimate:  214,887
#> Mean IBNR:       53,900
#> IBNR.S.E         18,925
#> Total IBNR 75%:  64,342
#> Total IBNR 95%:  88,540
summary(B)
#> $ByOrigin
#>      Latest Mean Ultimate  Mean IBNR    SD IBNR  IBNR 75%  IBNR 95%
#> 1981  18834      18834.00     0.0000     0.0000     0.000     0.000
#> 1982  16704      16859.68   155.6849   691.3395   187.548  1415.969
#> 1983  23466      24124.09   658.0872  1335.7521  1037.215  3332.854
#> 1984  27067      28760.97  1693.9667  1851.2270  2508.219  5441.105
#> 1985  26180      29016.35  2836.3546  2330.4607  4209.491  7128.393
#> 1986  15852      19646.51  3794.5072  2579.9212  5068.260  8810.207
#> 1987  12314      17649.63  5335.6303  3164.6874  6907.452 11092.871
#> 1988  13112      24207.75 11095.7478  5013.8173 13914.268 20546.085
#> 1989   5395      16231.48 10836.4756  6029.6321 14420.665 22058.061
#> 1990   2063      19556.69 17493.6861 13885.2032 24849.473 43075.450
#> 
#> $Totals
#>                    Totals
#> Latest:         160987.00
#> Mean Ultimate:  214887.14
#> Mean IBNR:       53900.14
#> SD IBNR:         18924.51
#> Total IBNR 75%:  64342.44
#> Total IBNR 95%:  88540.29
#> 
mean(B)
#> $ByOrigin
#>       Mean IBNR
#> 1981     0.0000
#> 1982   155.6849
#> 1983   658.0872
#> 1984  1693.9667
#> 1985  2836.3546
#> 1986  3794.5072
#> 1987  5335.6303
#> 1988 11095.7478
#> 1989 10836.4756
#> 1990 17493.6861
#> 
#> $Totals
#>               Total
#> Mean IBNR: 53900.14
#> 
quantile(B, c(0.75,0.95,0.99, 0.995))
#> $ByOrigin
#>       IBNR 75%  IBNR 95%  IBNR 99% IBNR 99.5%
#> 1981     0.000     0.000     0.000      0.000
#> 1982   187.548  1415.969  2774.821   3278.935
#> 1983  1037.215  3332.854  5398.659   6690.172
#> 1984  2508.219  5441.105  7398.436   8661.683
#> 1985  4209.491  7128.393  9893.090  11070.283
#> 1986  5068.260  8810.207 11307.528  13005.806
#> 1987  6907.452 11092.871 16625.553  17698.202
#> 1988 13914.268 20546.085 26582.477  27675.383
#> 1989 14420.665 22058.061 26912.534  28314.998
#> 1990 24849.473 43075.450 62276.411  63406.865
#> 
#> $Totals
#>                Totals
#> IBNR 75%:    64342.44
#> IBNR 95%:    88540.29
#> IBNR 99%:   106809.17
#> IBNR 99.5%: 120225.58
#>