Methods for BootChainLadder objects

summary, print, mean, and quantile methods for BootChainLadder objects

Usage

# S3 method for BootChainLadder
summary(object, probs=c(0.75,0.95), ...)

# S3 method for BootChainLadder
print(x, probs=c(0.75,0.95), ...)

# S3 method for BootChainLadder
quantile(x, probs=c(0.75, 0.95), na.rm = FALSE,
              names = TRUE, type = 7,...)

# S3 method for BootChainLadder
mean(x, ...)

# S3 method for BootChainLadder
residuals(object, ...)

Arguments

x, object

output from BootChainLadder

probs

numeric vector of probabilities with values in [0,1], see quantile for more help

na.rm

logical; if true, any NA and NaN's are removed from 'x' before the quantiles are computed, see quantile for more help

names

logical; if true, the result has a names attribute. Set to FALSE for speedup with many 'probs', see quantile for more help

type

an integer between 1 and 9 selecting one of the nine quantile algorithms detailed below to be used, see quantile

...

further arguments passed to or from other methods

Details

print.BootChainLadder calls summary.BootChainLadder and prints a formatted version of the summary. residuals.BootChainLadder gives the residual triangle of the expected chain-ladder minus the actual triangle back.

Value

summary.BootChainLadder, mean.BootChainLadder, and quantile.BootChainLadder, give a list with two elements back:

ByOrigin

data frame with summary/mean/quantile statistics by origin period

Totals

data frame with total summary/mean/quantile statistics for all origin period

Author

Markus Gesmann

See also

Examples

B <- BootChainLadder(RAA, R=999, process.distr="gamma")
B
#> BootChainLadder(Triangle = RAA, R = 999, process.distr = "gamma")
#> 
#>      Latest Mean Ultimate Mean IBNR IBNR.S.E IBNR 75% IBNR 95%
#> 1981 18,834        18,834         0        0        0        0
#> 1982 16,704        16,887       183      715      216    1,383
#> 1983 23,466        24,084       618    1,335    1,139    3,195
#> 1984 27,067        28,831     1,764    1,936    2,680    5,448
#> 1985 26,180        29,045     2,865    2,331    4,008    7,363
#> 1986 15,852        19,530     3,678    2,400    4,976    8,305
#> 1987 12,314        17,731     5,417    3,084    7,225   11,501
#> 1988 13,112        24,415    11,303    4,963   14,592   19,711
#> 1989  5,395        16,410    11,015    6,272   14,531   22,734
#> 1990  2,063        19,392    17,329   13,538   24,839   42,465
#> 
#>                  Totals
#> Latest:         160,987
#> Mean Ultimate:  215,157
#> Mean IBNR:       54,170
#> IBNR.S.E         19,273
#> Total IBNR 75%:  65,659
#> Total IBNR 95%:  90,866
summary(B)
#> $ByOrigin
#>      Latest Mean Ultimate  Mean IBNR   SD IBNR   IBNR 75%  IBNR 95%
#> 1981  18834      18834.00     0.0000     0.000     0.0000     0.000
#> 1982  16704      16886.61   182.6132   715.151   216.0255  1382.667
#> 1983  23466      24084.09   618.0918  1335.049  1139.2133  3195.060
#> 1984  27067      28830.77  1763.7690  1935.502  2680.2663  5448.308
#> 1985  26180      29045.37  2865.3725  2330.830  4007.9752  7362.615
#> 1986  15852      19529.76  3677.7623  2400.365  4975.8297  8304.909
#> 1987  12314      17730.86  5416.8643  3084.476  7224.9389 11501.482
#> 1988  13112      24414.57 11302.5682  4962.658 14591.9258 19710.544
#> 1989   5395      16409.69 11014.6925  6272.245 14530.9135 22733.648
#> 1990   2063      19391.60 17328.6040 13537.670 24838.9111 42465.481
#> 
#> $Totals
#>                    Totals
#> Latest:         160987.00
#> Mean Ultimate:  215157.34
#> Mean IBNR:       54170.34
#> SD IBNR:         19273.02
#> Total IBNR 75%:  65658.63
#> Total IBNR 95%:  90865.97
#> 
mean(B)
#> $ByOrigin
#>       Mean IBNR
#> 1981     0.0000
#> 1982   182.6132
#> 1983   618.0918
#> 1984  1763.7690
#> 1985  2865.3725
#> 1986  3677.7623
#> 1987  5416.8643
#> 1988 11302.5682
#> 1989 11014.6925
#> 1990 17328.6040
#> 
#> $Totals
#>               Total
#> Mean IBNR: 54170.34
#> 
quantile(B, c(0.75,0.95,0.99, 0.995))
#> $ByOrigin
#>        IBNR 75%  IBNR 95%  IBNR 99% IBNR 99.5%
#> 1981     0.0000     0.000     0.000      0.000
#> 1982   216.0255  1382.667  3108.816   3193.868
#> 1983  1139.2133  3195.060  4803.709   5892.061
#> 1984  2680.2663  5448.308  7688.048   8559.684
#> 1985  4007.9752  7362.615  9774.725  11059.348
#> 1986  4975.8297  8304.909 10980.727  12101.399
#> 1987  7224.9389 11501.482 13509.884  15142.547
#> 1988 14591.9258 19710.544 24384.018  25841.382
#> 1989 14530.9135 22733.648 28151.098  31162.739
#> 1990 24838.9111 42465.481 57363.280  60889.192
#> 
#> $Totals
#>                Totals
#> IBNR 75%:    65658.63
#> IBNR 95%:    90865.97
#> IBNR 99%:   102577.14
#> IBNR 99.5%: 106082.80
#>