Methods for BootChainLadder objects

summary, print, mean, and quantile methods for BootChainLadder objects

Usage

# S3 method for BootChainLadder
summary(object, probs=c(0.75,0.95), ...)

# S3 method for BootChainLadder
print(x, probs=c(0.75,0.95), ...)

# S3 method for BootChainLadder
quantile(x, probs=c(0.75, 0.95), na.rm = FALSE,
              names = TRUE, type = 7,...)

# S3 method for BootChainLadder
mean(x, ...)

# S3 method for BootChainLadder
residuals(object, ...)

Arguments

x, object

output from BootChainLadder

probs

numeric vector of probabilities with values in [0,1], see quantile for more help

na.rm

logical; if true, any NA and NaN's are removed from 'x' before the quantiles are computed, see quantile for more help

names

logical; if true, the result has a names attribute. Set to FALSE for speedup with many 'probs', see quantile for more help

type

an integer between 1 and 9 selecting one of the nine quantile algorithms detailed below to be used, see quantile

...

further arguments passed to or from other methods

Details

print.BootChainLadder calls summary.BootChainLadder and prints a formatted version of the summary. residuals.BootChainLadder gives the residual triangle of the expected chain-ladder minus the actual triangle back.

Value

summary.BootChainLadder, mean.BootChainLadder, and

quantile.BootChainLadder, give a list with two elements back:

ByOrigin

data frame with summary/mean/quantile statistics by origin period

Totals

data frame with total summary/mean/quantile statistics for all origin period

Author

Markus Gesmann

See also

Examples

B <- BootChainLadder(RAA, R=999, process.distr="gamma")
B
#> BootChainLadder(Triangle = RAA, R = 999, process.distr = "gamma")
#> 
#>      Latest Mean Ultimate Mean IBNR IBNR.S.E IBNR 75% IBNR 95%
#> 1981 18,834        18,834         0        0        0        0
#> 1982 16,704        16,860       156      774      166    1,404
#> 1983 23,466        24,123       657    1,324    1,083    3,084
#> 1984 27,067        28,796     1,729    1,872    2,680    5,397
#> 1985 26,180        28,955     2,775    2,254    3,974    6,871
#> 1986 15,852        19,513     3,661    2,439    5,014    8,149
#> 1987 12,314        18,019     5,705    3,331    7,616   11,633
#> 1988 13,112        24,365    11,253    5,246   14,441   21,114
#> 1989  5,395        16,086    10,691    6,099   14,373   22,187
#> 1990  2,063        19,582    17,519   13,571   25,450   42,647
#> 
#>                  Totals
#> Latest:         160,987
#> Mean Ultimate:  215,133
#> Mean IBNR:       54,146
#> IBNR.S.E         18,977
#> Total IBNR 75%:  65,745
#> Total IBNR 95%:  88,517
summary(B)
#> $ByOrigin
#>      Latest Mean Ultimate  Mean IBNR    SD IBNR   IBNR 75%  IBNR 95%
#> 1981  18834      18834.00     0.0000     0.0000     0.0000     0.000
#> 1982  16704      16859.83   155.8285   774.3399   166.3734  1403.757
#> 1983  23466      24123.29   657.2875  1323.9797  1082.6877  3083.735
#> 1984  27067      28795.82  1728.8172  1871.8530  2679.6504  5396.742
#> 1985  26180      28954.68  2774.6755  2254.0021  3974.2965  6871.064
#> 1986  15852      19512.81  3660.8086  2439.3744  5013.5928  8149.085
#> 1987  12314      18018.78  5704.7755  3331.3257  7615.6818 11633.225
#> 1988  13112      24364.99 11252.9884  5245.6875 14441.3123 21114.302
#> 1989   5395      16086.03 10691.0289  6098.8642 14372.9742 22187.125
#> 1990   2063      19582.48 17519.4750 13571.3229 25449.5355 42646.852
#> 
#> $Totals
#>                    Totals
#> Latest:         160987.00
#> Mean Ultimate:  215132.69
#> Mean IBNR:       54145.69
#> SD IBNR:         18977.12
#> Total IBNR 75%:  65745.47
#> Total IBNR 95%:  88517.06
#> 
mean(B)
#> $ByOrigin
#>       Mean IBNR
#> 1981     0.0000
#> 1982   155.8285
#> 1983   657.2875
#> 1984  1728.8172
#> 1985  2774.6755
#> 1986  3660.8086
#> 1987  5704.7755
#> 1988 11252.9884
#> 1989 10691.0289
#> 1990 17519.4750
#> 
#> $Totals
#>               Total
#> Mean IBNR: 54145.69
#> 
quantile(B, c(0.75,0.95,0.99, 0.995))
#> $ByOrigin
#>        IBNR 75%  IBNR 95%  IBNR 99% IBNR 99.5%
#> 1981     0.0000     0.000     0.000      0.000
#> 1982   166.3734  1403.757  2907.167   3648.016
#> 1983  1082.6877  3083.735  5546.293   6372.405
#> 1984  2679.6504  5396.742  7395.160   7956.302
#> 1985  3974.2965  6871.064  9430.694  10037.353
#> 1986  5013.5928  8149.085 11736.293  12436.247
#> 1987  7615.6818 11633.225 15211.437  17201.604
#> 1988 14441.3123 21114.302 25684.952  27355.092
#> 1989 14372.9742 22187.125 28526.534  30959.823
#> 1990 25449.5355 42646.852 58968.385  65002.912
#> 
#> $Totals
#>                Totals
#> IBNR 75%:    65745.47
#> IBNR 95%:    88517.06
#> IBNR 99%:   105585.60
#> IBNR 99.5%: 111729.94
#>