Methods for BootChainLadder objects

summary, print, mean, and quantile methods for BootChainLadder objects

Usage

# S3 method for BootChainLadder
summary(object, probs=c(0.75,0.95), ...)

# S3 method for BootChainLadder
print(x, probs=c(0.75,0.95), ...)

# S3 method for BootChainLadder
quantile(x, probs=c(0.75, 0.95), na.rm = FALSE,
              names = TRUE, type = 7,...)

# S3 method for BootChainLadder
mean(x, ...)

# S3 method for BootChainLadder
residuals(object, ...)

Arguments

x, object

output from BootChainLadder

probs

numeric vector of probabilities with values in [0,1], see quantile for more help

na.rm

logical; if true, any NA and NaN's are removed from 'x' before the quantiles are computed, see quantile for more help

names

logical; if true, the result has a names attribute. Set to FALSE for speedup with many 'probs', see quantile for more help

type

an integer between 1 and 9 selecting one of the nine quantile algorithms detailed below to be used, see quantile

...

further arguments passed to or from other methods

Details

print.BootChainLadder calls summary.BootChainLadder and prints a formatted version of the summary. residuals.BootChainLadder gives the residual triangle of the expected chain-ladder minus the actual triangle back.

Value

summary.BootChainLadder, mean.BootChainLadder, and

quantile.BootChainLadder, give a list with two elements back:

ByOrigin

data frame with summary/mean/quantile statistics by origin period

Totals

data frame with total summary/mean/quantile statistics for all origin period

Author

Markus Gesmann

See also

Examples

B <- BootChainLadder(RAA, R=999, process.distr="gamma")
B
#> BootChainLadder(Triangle = RAA, R = 999, process.distr = "gamma")
#> 
#>      Latest Mean Ultimate Mean IBNR IBNR.S.E IBNR 75% IBNR 95%
#> 1981 18,834        18,834         0        0        0        0
#> 1982 16,704        16,881       177      686      189    1,397
#> 1983 23,466        24,146       680    1,282    1,191    3,240
#> 1984 27,067        28,696     1,629    1,838    2,635    4,813
#> 1985 26,180        28,964     2,784    2,356    4,014    6,865
#> 1986 15,852        19,720     3,868    2,590    5,278    8,842
#> 1987 12,314        17,804     5,490    3,097    7,194   11,193
#> 1988 13,112        24,117    11,005    4,846   13,720   20,150
#> 1989  5,395        16,581    11,186    5,852   14,965   21,230
#> 1990  2,063        19,581    17,518   13,649   25,762   41,915
#> 
#>                  Totals
#> Latest:         160,987
#> Mean Ultimate:  215,324
#> Mean IBNR:       54,337
#> IBNR.S.E         18,695
#> Total IBNR 75%:  66,336
#> Total IBNR 95%:  88,368
summary(B)
#> $ByOrigin
#>      Latest Mean Ultimate  Mean IBNR    SD IBNR   IBNR 75%  IBNR 95%
#> 1981  18834      18834.00     0.0000     0.0000     0.0000     0.000
#> 1982  16704      16881.09   177.0861   686.2623   189.0542  1397.154
#> 1983  23466      24146.11   680.1094  1282.2632  1191.4152  3240.174
#> 1984  27067      28696.26  1629.2560  1837.8909  2634.6144  4813.323
#> 1985  26180      28963.60  2783.5969  2355.7155  4014.2190  6865.141
#> 1986  15852      19720.30  3868.3009  2590.1606  5278.3533  8841.978
#> 1987  12314      17803.50  5489.5032  3097.0534  7194.1471 11193.283
#> 1988  13112      24116.51 11004.5115  4846.1596 13720.2724 20150.044
#> 1989   5395      16581.29 11186.2924  5852.0470 14964.6030 21229.538
#> 1990   2063      19580.88 17517.8785 13648.5910 25761.8322 41915.346
#> 
#> $Totals
#>                    Totals
#> Latest:         160987.00
#> Mean Ultimate:  215323.53
#> Mean IBNR:       54336.53
#> SD IBNR:         18694.67
#> Total IBNR 75%:  66335.94
#> Total IBNR 95%:  88367.59
#> 
mean(B)
#> $ByOrigin
#>       Mean IBNR
#> 1981     0.0000
#> 1982   177.0861
#> 1983   680.1094
#> 1984  1629.2560
#> 1985  2783.5969
#> 1986  3868.3009
#> 1987  5489.5032
#> 1988 11004.5115
#> 1989 11186.2924
#> 1990 17517.8785
#> 
#> $Totals
#>               Total
#> Mean IBNR: 54336.53
#> 
quantile(B, c(0.75,0.95,0.99, 0.995))
#> $ByOrigin
#>        IBNR 75%  IBNR 95%  IBNR 99% IBNR 99.5%
#> 1981     0.0000     0.000     0.000      0.000
#> 1982   189.0542  1397.154  3020.420   3281.758
#> 1983  1191.4152  3240.174  4836.908   5974.986
#> 1984  2634.6144  4813.323  6921.692   8299.420
#> 1985  4014.2190  6865.141 10104.071  12354.328
#> 1986  5278.3533  8841.978 11216.793  11646.767
#> 1987  7194.1471 11193.283 14982.144  16792.575
#> 1988 13720.2724 20150.044 23957.905  27208.110
#> 1989 14964.6030 21229.538 26894.336  30885.700
#> 1990 25761.8322 41915.346 56532.962  63656.976
#> 
#> $Totals
#>                Totals
#> IBNR 75%:    66335.94
#> IBNR 95%:    88367.59
#> IBNR 99%:   104841.16
#> IBNR 99.5%: 108577.84
#>