Methods for BootChainLadder objects

summary, print, mean, and quantile methods for BootChainLadder objects

Usage

# S3 method for BootChainLadder
summary(object, probs=c(0.75,0.95), ...)

# S3 method for BootChainLadder
print(x, probs=c(0.75,0.95), ...)

# S3 method for BootChainLadder
quantile(x, probs=c(0.75, 0.95), na.rm = FALSE,
              names = TRUE, type = 7,...)

# S3 method for BootChainLadder
mean(x, ...)

# S3 method for BootChainLadder
residuals(object, ...)

Arguments

x, object

output from BootChainLadder

probs

numeric vector of probabilities with values in [0,1], see quantile for more help

na.rm

logical; if true, any NA and NaN's are removed from 'x' before the quantiles are computed, see quantile for more help

names

logical; if true, the result has a names attribute. Set to FALSE for speedup with many 'probs', see quantile for more help

type

an integer between 1 and 9 selecting one of the nine quantile algorithms detailed below to be used, see quantile

...

further arguments passed to or from other methods

Details

print.BootChainLadder calls summary.BootChainLadder and prints a formatted version of the summary. residuals.BootChainLadder gives the residual triangle of the expected chain-ladder minus the actual triangle back.

Value

summary.BootChainLadder, mean.BootChainLadder, and

quantile.BootChainLadder, give a list with two elements back:

ByOrigin

data frame with summary/mean/quantile statistics by origin period

Totals

data frame with total summary/mean/quantile statistics for all origin period

Author

Markus Gesmann

See also

Examples

B <- BootChainLadder(RAA, R=999, process.distr="gamma")
B
#> BootChainLadder(Triangle = RAA, R = 999, process.distr = "gamma")
#> 
#>      Latest Mean Ultimate Mean IBNR IBNR.S.E IBNR 75% IBNR 95%
#> 1981 18,834        18,834         0        0        0        0
#> 1982 16,704        16,882       178      726      223    1,316
#> 1983 23,466        24,178       712    1,303    1,185    3,300
#> 1984 27,067        28,822     1,755    1,881    2,753    5,252
#> 1985 26,180        29,003     2,823    2,294    3,971    7,278
#> 1986 15,852        19,527     3,675    2,427    4,953    8,276
#> 1987 12,314        18,019     5,705    3,238    7,579   11,876
#> 1988 13,112        23,953    10,841    4,792   13,597   19,613
#> 1989  5,395        16,420    11,025    5,954   14,541   21,478
#> 1990  2,063        19,998    17,935   14,060   24,922   44,390
#> 
#>                  Totals
#> Latest:         160,987
#> Mean Ultimate:  215,637
#> Mean IBNR:       54,650
#> IBNR.S.E         18,286
#> Total IBNR 75%:  65,222
#> Total IBNR 95%:  87,806
summary(B)
#> $ByOrigin
#>      Latest Mean Ultimate  Mean IBNR    SD IBNR   IBNR 75%  IBNR 95%
#> 1981  18834      18834.00     0.0000     0.0000     0.0000     0.000
#> 1982  16704      16881.96   177.9623   725.5492   222.6855  1315.546
#> 1983  23466      24178.35   712.3533  1303.1483  1184.8490  3300.252
#> 1984  27067      28822.45  1755.4490  1880.9638  2753.4133  5252.071
#> 1985  26180      29002.91  2822.9090  2294.3745  3971.1315  7277.564
#> 1986  15852      19526.84  3674.8444  2426.6009  4953.4033  8275.875
#> 1987  12314      18019.26  5705.2612  3238.0394  7579.0081 11875.939
#> 1988  13112      23952.84 10840.8415  4791.8286 13596.6636 19612.819
#> 1989   5395      16419.77 11024.7728  5953.5399 14541.4754 21478.387
#> 1990   2063      19998.49 17935.4892 14060.0262 24922.4742 44390.211
#> 
#> $Totals
#>                    Totals
#> Latest:         160987.00
#> Mean Ultimate:  215636.88
#> Mean IBNR:       54649.88
#> SD IBNR:         18286.45
#> Total IBNR 75%:  65222.03
#> Total IBNR 95%:  87806.34
#> 
mean(B)
#> $ByOrigin
#>       Mean IBNR
#> 1981     0.0000
#> 1982   177.9623
#> 1983   712.3533
#> 1984  1755.4490
#> 1985  2822.9090
#> 1986  3674.8444
#> 1987  5705.2612
#> 1988 10840.8415
#> 1989 11024.7728
#> 1990 17935.4892
#> 
#> $Totals
#>               Total
#> Mean IBNR: 54649.88
#> 
quantile(B, c(0.75,0.95,0.99, 0.995))
#> $ByOrigin
#>        IBNR 75%  IBNR 95%  IBNR 99% IBNR 99.5%
#> 1981     0.0000     0.000     0.000      0.000
#> 1982   222.6855  1315.546  3213.533   3839.519
#> 1983  1184.8490  3300.252  4951.342   5475.705
#> 1984  2753.4133  5252.071  6962.023   8157.257
#> 1985  3971.1315  7277.564  9697.627  11432.862
#> 1986  4953.4033  8275.875 11655.427  12571.126
#> 1987  7579.0081 11875.939 15133.584  16489.104
#> 1988 13596.6636 19612.819 25074.109  26504.964
#> 1989 14541.4754 21478.387 28438.307  30563.599
#> 1990 24922.4742 44390.211 57932.847  62783.818
#> 
#> $Totals
#>                Totals
#> IBNR 75%:    65222.03
#> IBNR 95%:    87806.34
#> IBNR 99%:   102406.64
#> IBNR 99.5%: 110946.32
#>