Methods for BootChainLadder objects

summary, print, mean, and quantile methods for BootChainLadder objects

Usage

# S3 method for BootChainLadder
summary(object, probs=c(0.75,0.95), ...)

# S3 method for BootChainLadder
print(x, probs=c(0.75,0.95), ...)

# S3 method for BootChainLadder
quantile(x, probs=c(0.75, 0.95), na.rm = FALSE,
              names = TRUE, type = 7,...)

# S3 method for BootChainLadder
mean(x, ...)

# S3 method for BootChainLadder
residuals(object, ...)

Arguments

x, object

output from BootChainLadder

probs

numeric vector of probabilities with values in [0,1], see quantile for more help

na.rm

logical; if true, any NA and NaN's are removed from 'x' before the quantiles are computed, see quantile for more help

names

logical; if true, the result has a names attribute. Set to FALSE for speedup with many 'probs', see quantile for more help

type

an integer between 1 and 9 selecting one of the nine quantile algorithms detailed below to be used, see quantile

...

further arguments passed to or from other methods

Details

print.BootChainLadder calls summary.BootChainLadder and prints a formatted version of the summary. residuals.BootChainLadder gives the residual triangle of the expected chain-ladder minus the actual triangle back.

Value

summary.BootChainLadder, mean.BootChainLadder, and quantile.BootChainLadder, give a list with two elements back:

ByOrigin

data frame with summary/mean/quantile statistics by origin period

Totals

data frame with total summary/mean/quantile statistics for all origin period

Author

Markus Gesmann

See also

Examples

B <- BootChainLadder(RAA, R=999, process.distr="gamma")
B
#> BootChainLadder(Triangle = RAA, R = 999, process.distr = "gamma")
#> 
#>      Latest Mean Ultimate Mean IBNR IBNR.S.E IBNR 75% IBNR 95%
#> 1981 18,834        18,834         0        0        0        0
#> 1982 16,704        16,888       184      687      195    1,385
#> 1983 23,466        24,150       684    1,383    1,255    3,215
#> 1984 27,067        28,732     1,665    1,841    2,516    5,337
#> 1985 26,180        29,054     2,874    2,329    4,107    7,136
#> 1986 15,852        19,674     3,822    2,609    5,154    8,528
#> 1987 12,314        17,787     5,473    2,959    7,303   10,856
#> 1988 13,112        24,259    11,147    5,087   14,116   20,678
#> 1989  5,395        16,313    10,918    6,169   14,679   22,257
#> 1990  2,063        18,958    16,895   13,501   24,673   41,477
#> 
#>                  Totals
#> Latest:         160,987
#> Mean Ultimate:  214,648
#> Mean IBNR:       53,661
#> IBNR.S.E         18,777
#> Total IBNR 75%:  64,885
#> Total IBNR 95%:  88,019
summary(B)
#> $ByOrigin
#>      Latest Mean Ultimate  Mean IBNR    SD IBNR   IBNR 75%  IBNR 95%
#> 1981  18834      18834.00     0.0000     0.0000     0.0000     0.000
#> 1982  16704      16887.54   183.5363   686.8667   195.0937  1384.616
#> 1983  23466      24150.09   684.0913  1383.4998  1255.0542  3214.774
#> 1984  27067      28731.88  1664.8794  1841.1892  2515.9076  5336.877
#> 1985  26180      29053.53  2873.5260  2329.1302  4107.4308  7136.340
#> 1986  15852      19674.45  3822.4487  2609.4455  5153.9737  8528.025
#> 1987  12314      17786.95  5472.9541  2958.5637  7303.4793 10856.474
#> 1988  13112      24258.68 11146.6833  5086.8890 14116.4593 20677.791
#> 1989   5395      16312.79 10917.7917  6169.4981 14679.2350 22256.530
#> 1990   2063      18957.73 16894.7270 13500.8210 24673.4849 41477.475
#> 
#> $Totals
#>                    Totals
#> Latest:         160987.00
#> Mean Ultimate:  214647.64
#> Mean IBNR:       53660.64
#> SD IBNR:         18776.75
#> Total IBNR 75%:  64884.72
#> Total IBNR 95%:  88019.16
#> 
mean(B)
#> $ByOrigin
#>       Mean IBNR
#> 1981     0.0000
#> 1982   183.5363
#> 1983   684.0913
#> 1984  1664.8794
#> 1985  2873.5260
#> 1986  3822.4487
#> 1987  5472.9541
#> 1988 11146.6833
#> 1989 10917.7917
#> 1990 16894.7270
#> 
#> $Totals
#>               Total
#> Mean IBNR: 53660.64
#> 
quantile(B, c(0.75,0.95,0.99, 0.995))
#> $ByOrigin
#>        IBNR 75%  IBNR 95%  IBNR 99% IBNR 99.5%
#> 1981     0.0000     0.000     0.000      0.000
#> 1982   195.0937  1384.616  2737.434   3138.534
#> 1983  1255.0542  3214.774  4935.900   5775.627
#> 1984  2515.9076  5336.877  7623.917   8412.321
#> 1985  4107.4308  7136.340  9611.322  10503.218
#> 1986  5153.9737  8528.025 12493.001  13967.455
#> 1987  7303.4793 10856.474 13839.821  14745.969
#> 1988 14116.4593 20677.791 25632.589  27545.035
#> 1989 14679.2350 22256.530 28022.468  31497.797
#> 1990 24673.4849 41477.475 53872.669  55913.925
#> 
#> $Totals
#>                Totals
#> IBNR 75%:    64884.72
#> IBNR 95%:    88019.16
#> IBNR 99%:   101753.70
#> IBNR 99.5%: 106097.39
#>