Methods for BootChainLadder objects

summary, print, mean, and quantile methods for BootChainLadder objects

Usage

# S3 method for BootChainLadder
summary(object, probs=c(0.75,0.95), ...)

# S3 method for BootChainLadder
print(x, probs=c(0.75,0.95), ...)

# S3 method for BootChainLadder
quantile(x, probs=c(0.75, 0.95), na.rm = FALSE,
              names = TRUE, type = 7,...)

# S3 method for BootChainLadder
mean(x, ...)

# S3 method for BootChainLadder
residuals(object, ...)

Arguments

x, object

output from BootChainLadder

probs

numeric vector of probabilities with values in [0,1], see quantile for more help

na.rm

logical; if true, any NA and NaN's are removed from 'x' before the quantiles are computed, see quantile for more help

names

logical; if true, the result has a names attribute. Set to FALSE for speedup with many 'probs', see quantile for more help

type

an integer between 1 and 9 selecting one of the nine quantile algorithms detailed below to be used, see quantile

...

further arguments passed to or from other methods

Details

print.BootChainLadder calls summary.BootChainLadder and prints a formatted version of the summary. residuals.BootChainLadder gives the residual triangle of the expected chain-ladder minus the actual triangle back.

Value

summary.BootChainLadder, mean.BootChainLadder, and

quantile.BootChainLadder, give a list with two elements back:

ByOrigin

data frame with summary/mean/quantile statistics by origin period

Totals

data frame with total summary/mean/quantile statistics for all origin period

Author

Markus Gesmann

See also

Examples

B <- BootChainLadder(RAA, R=999, process.distr="gamma")
B
#> BootChainLadder(Triangle = RAA, R = 999, process.distr = "gamma")
#> 
#>      Latest Mean Ultimate Mean IBNR IBNR.S.E IBNR 75% IBNR 95%
#> 1981 18,834        18,834         0        0        0        0
#> 1982 16,704        16,839       135      665      180    1,307
#> 1983 23,466        24,076       610    1,271    1,163    3,061
#> 1984 27,067        28,684     1,617    1,903    2,565    5,271
#> 1985 26,180        28,948     2,768    2,265    4,064    6,681
#> 1986 15,852        19,633     3,781    2,597    5,210    8,765
#> 1987 12,314        17,664     5,350    2,917    6,853   10,936
#> 1988 13,112        24,157    11,045    4,926   13,717   20,451
#> 1989  5,395        16,152    10,757    6,222   14,307   22,202
#> 1990  2,063        19,203    17,140   13,228   24,331   40,117
#> 
#>                  Totals
#> Latest:         160,987
#> Mean Ultimate:  214,191
#> Mean IBNR:       53,204
#> IBNR.S.E         18,373
#> Total IBNR 75%:  64,339
#> Total IBNR 95%:  86,740
summary(B)
#> $ByOrigin
#>      Latest Mean Ultimate  Mean IBNR    SD IBNR   IBNR 75%  IBNR 95%
#> 1981  18834      18834.00     0.0000     0.0000     0.0000     0.000
#> 1982  16704      16839.04   135.0409   664.6513   179.6239  1307.276
#> 1983  23466      24076.34   610.3380  1270.5197  1163.2053  3061.025
#> 1984  27067      28683.52  1616.5214  1902.6642  2564.8094  5271.011
#> 1985  26180      28948.27  2768.2739  2264.7189  4064.3354  6681.140
#> 1986  15852      19632.71  3780.7108  2597.4651  5209.5445  8764.773
#> 1987  12314      17664.35  5350.3486  2917.2805  6852.6106 10935.792
#> 1988  13112      24157.29 11045.2855  4926.3111 13716.7720 20450.890
#> 1989   5395      16151.74 10756.7406  6221.5195 14307.3589 22202.083
#> 1990   2063      19203.49 17140.4882 13227.7013 24330.9399 40117.065
#> 
#> $Totals
#>                    Totals
#> Latest:         160987.00
#> Mean Ultimate:  214190.75
#> Mean IBNR:       53203.75
#> SD IBNR:         18372.86
#> Total IBNR 75%:  64338.78
#> Total IBNR 95%:  86739.65
#> 
mean(B)
#> $ByOrigin
#>       Mean IBNR
#> 1981     0.0000
#> 1982   135.0409
#> 1983   610.3380
#> 1984  1616.5214
#> 1985  2768.2739
#> 1986  3780.7108
#> 1987  5350.3486
#> 1988 11045.2855
#> 1989 10756.7406
#> 1990 17140.4882
#> 
#> $Totals
#>               Total
#> Mean IBNR: 53203.75
#> 
quantile(B, c(0.75,0.95,0.99, 0.995))
#> $ByOrigin
#>        IBNR 75%  IBNR 95%  IBNR 99% IBNR 99.5%
#> 1981     0.0000     0.000     0.000      0.000
#> 1982   179.6239  1307.276  2612.015   3649.232
#> 1983  1163.2053  3061.025  4540.335   5219.247
#> 1984  2564.8094  5271.011  6994.063   7584.374
#> 1985  4064.3354  6681.140  9232.779   9961.691
#> 1986  5209.5445  8764.773 11280.364  12854.059
#> 1987  6852.6106 10935.792 13828.889  15418.082
#> 1988 13716.7720 20450.890 24900.688  26140.267
#> 1989 14307.3589 22202.083 27975.089  33491.243
#> 1990 24330.9399 40117.065 58574.983  66203.984
#> 
#> $Totals
#>                Totals
#> IBNR 75%:    64338.78
#> IBNR 95%:    86739.65
#> IBNR 99%:   104498.56
#> IBNR 99.5%: 107606.41
#>