Methods for BootChainLadder objects

summary, print, mean, and quantile methods for BootChainLadder objects

Usage

# S3 method for BootChainLadder
summary(object, probs=c(0.75,0.95), ...)

# S3 method for BootChainLadder
print(x, probs=c(0.75,0.95), ...)

# S3 method for BootChainLadder
quantile(x, probs=c(0.75, 0.95), na.rm = FALSE,
              names = TRUE, type = 7,...)

# S3 method for BootChainLadder
mean(x, ...)

# S3 method for BootChainLadder
residuals(object, ...)

Arguments

x, object

output from BootChainLadder

probs

numeric vector of probabilities with values in [0,1], see quantile for more help

na.rm

logical; if true, any NA and NaN's are removed from 'x' before the quantiles are computed, see quantile for more help

names

logical; if true, the result has a names attribute. Set to FALSE for speedup with many 'probs', see quantile for more help

type

an integer between 1 and 9 selecting one of the nine quantile algorithms detailed below to be used, see quantile

...

further arguments passed to or from other methods

Details

print.BootChainLadder calls summary.BootChainLadder and prints a formatted version of the summary. residuals.BootChainLadder gives the residual triangle of the expected chain-ladder minus the actual triangle back.

Value

summary.BootChainLadder, mean.BootChainLadder, and

quantile.BootChainLadder, give a list with two elements back:

ByOrigin

data frame with summary/mean/quantile statistics by origin period

Totals

data frame with total summary/mean/quantile statistics for all origin period

Author

Markus Gesmann

See also

Examples

B <- BootChainLadder(RAA, R=999, process.distr="gamma")
B
#> BootChainLadder(Triangle = RAA, R = 999, process.distr = "gamma")
#> 
#>      Latest Mean Ultimate Mean IBNR IBNR.S.E IBNR 75% IBNR 95%
#> 1981 18,834        18,834         0        0        0        0
#> 1982 16,704        16,840       136      663      164    1,332
#> 1983 23,466        24,022       556    1,343    1,028    3,052
#> 1984 27,067        28,755     1,688    1,870    2,650    5,328
#> 1985 26,180        28,885     2,705    2,356    3,949    7,212
#> 1986 15,852        19,513     3,661    2,379    4,980    8,102
#> 1987 12,314        17,978     5,664    3,170    7,430   11,695
#> 1988 13,112        24,327    11,215    4,864   13,818   20,295
#> 1989  5,395        16,229    10,834    6,181   14,319   21,812
#> 1990  2,063        19,266    17,203   13,920   25,184   41,698
#> 
#>                  Totals
#> Latest:         160,987
#> Mean Ultimate:  214,649
#> Mean IBNR:       53,662
#> IBNR.S.E         18,849
#> Total IBNR 75%:  64,701
#> Total IBNR 95%:  85,776
summary(B)
#> $ByOrigin
#>      Latest Mean Ultimate  Mean IBNR    SD IBNR   IBNR 75%  IBNR 95%
#> 1981  18834      18834.00     0.0000     0.0000     0.0000     0.000
#> 1982  16704      16840.43   136.4281   663.3554   164.0136  1331.645
#> 1983  23466      24021.68   555.6790  1342.6910  1027.7946  3052.183
#> 1984  27067      28755.42  1688.4151  1870.2174  2649.5219  5327.796
#> 1985  26180      28885.44  2705.4403  2355.8833  3949.1886  7211.773
#> 1986  15852      19513.09  3661.0947  2379.1204  4979.8964  8101.802
#> 1987  12314      17977.68  5663.6793  3170.2039  7429.9975 11695.372
#> 1988  13112      24326.75 11214.7494  4864.4000 13817.7512 20295.313
#> 1989   5395      16229.29 10834.2885  6181.2602 14318.9433 21811.890
#> 1990   2063      19265.52 17202.5235 13919.6135 25184.3051 41697.942
#> 
#> $Totals
#>                    Totals
#> Latest:         160987.00
#> Mean Ultimate:  214649.30
#> Mean IBNR:       53662.30
#> SD IBNR:         18848.99
#> Total IBNR 75%:  64701.27
#> Total IBNR 95%:  85776.49
#> 
mean(B)
#> $ByOrigin
#>       Mean IBNR
#> 1981     0.0000
#> 1982   136.4281
#> 1983   555.6790
#> 1984  1688.4151
#> 1985  2705.4403
#> 1986  3661.0947
#> 1987  5663.6793
#> 1988 11214.7494
#> 1989 10834.2885
#> 1990 17202.5235
#> 
#> $Totals
#>              Total
#> Mean IBNR: 53662.3
#> 
quantile(B, c(0.75,0.95,0.99, 0.995))
#> $ByOrigin
#>        IBNR 75%  IBNR 95%  IBNR 99% IBNR 99.5%
#> 1981     0.0000     0.000     0.000      0.000
#> 1982   164.0136  1331.645  2382.465   3336.828
#> 1983  1027.7946  3052.183  5074.941   6497.123
#> 1984  2649.5219  5327.796  7860.890   8474.739
#> 1985  3949.1886  7211.773  9466.861  10083.665
#> 1986  4979.8964  8101.802 10975.775  11605.980
#> 1987  7429.9975 11695.372 14829.831  15371.683
#> 1988 13817.7512 20295.313 24617.597  25941.233
#> 1989 14318.9433 21811.890 29729.016  32598.480
#> 1990 25184.3051 41697.942 61187.414  67853.123
#> 
#> $Totals
#>                Totals
#> IBNR 75%:    64701.27
#> IBNR 95%:    85776.49
#> IBNR 99%:   111871.41
#> IBNR 99.5%: 122021.13
#>