Methods for BootChainLadder objects

summary, print, mean, and quantile methods for BootChainLadder objects

Usage

# S3 method for BootChainLadder
summary(object, probs=c(0.75,0.95), ...)

# S3 method for BootChainLadder
print(x, probs=c(0.75,0.95), ...)

# S3 method for BootChainLadder
quantile(x, probs=c(0.75, 0.95), na.rm = FALSE,
              names = TRUE, type = 7,...)

# S3 method for BootChainLadder
mean(x, ...)

# S3 method for BootChainLadder
residuals(object, ...)

Arguments

x, object

output from BootChainLadder

probs

numeric vector of probabilities with values in [0,1], see quantile for more help

na.rm

logical; if true, any NA and NaN's are removed from 'x' before the quantiles are computed, see quantile for more help

names

logical; if true, the result has a names attribute. Set to FALSE for speedup with many 'probs', see quantile for more help

type

an integer between 1 and 9 selecting one of the nine quantile algorithms detailed below to be used, see quantile

...

further arguments passed to or from other methods

Details

print.BootChainLadder calls summary.BootChainLadder and prints a formatted version of the summary. residuals.BootChainLadder gives the residual triangle of the expected chain-ladder minus the actual triangle back.

Value

summary.BootChainLadder, mean.BootChainLadder, and quantile.BootChainLadder, give a list with two elements back:

ByOrigin

data frame with summary/mean/quantile statistics by origin period

Totals

data frame with total summary/mean/quantile statistics for all origin period

Author

Markus Gesmann

See also

Examples

B <- BootChainLadder(RAA, R=999, process.distr="gamma")
B
#> BootChainLadder(Triangle = RAA, R = 999, process.distr = "gamma")
#> 
#>      Latest Mean Ultimate Mean IBNR IBNR.S.E IBNR 75% IBNR 95%
#> 1981 18,834        18,834         0        0        0        0
#> 1982 16,704        16,883       179      708      184    1,456
#> 1983 23,466        24,160       694    1,314    1,185    3,028
#> 1984 27,067        28,835     1,768    1,957    2,744    5,260
#> 1985 26,180        29,056     2,876    2,260    4,137    6,977
#> 1986 15,852        19,611     3,759    2,486    5,157    8,408
#> 1987 12,314        17,934     5,620    3,054    7,324   11,375
#> 1988 13,112        24,329    11,217    5,046   14,165   20,551
#> 1989  5,395        16,828    11,433    6,181   15,208   23,035
#> 1990  2,063        20,068    18,005   14,101   25,145   45,410
#> 
#>                  Totals
#> Latest:         160,987
#> Mean Ultimate:  216,538
#> Mean IBNR:       55,551
#> IBNR.S.E         19,208
#> Total IBNR 75%:  66,994
#> Total IBNR 95%:  91,530
summary(B)
#> $ByOrigin
#>      Latest Mean Ultimate  Mean IBNR    SD IBNR   IBNR 75%  IBNR 95%
#> 1981  18834      18834.00     0.0000     0.0000     0.0000     0.000
#> 1982  16704      16882.59   178.5882   707.9742   184.2002  1456.344
#> 1983  23466      24160.44   694.4367  1313.7458  1184.7417  3028.437
#> 1984  27067      28835.22  1768.2187  1957.2640  2744.4657  5260.209
#> 1985  26180      29055.59  2875.5878  2259.7506  4137.0464  6977.161
#> 1986  15852      19611.15  3759.1547  2485.6474  5156.7745  8407.995
#> 1987  12314      17934.23  5620.2341  3054.0600  7324.1357 11375.187
#> 1988  13112      24329.07 11217.0691  5046.4786 14164.8761 20551.177
#> 1989   5395      16827.71 11432.7087  6180.8871 15208.0374 23034.658
#> 1990   2063      20067.97 18004.9719 14100.8895 25144.7684 45410.332
#> 
#> $Totals
#>                    Totals
#> Latest:         160987.00
#> Mean Ultimate:  216537.97
#> Mean IBNR:       55550.97
#> SD IBNR:         19208.36
#> Total IBNR 75%:  66993.50
#> Total IBNR 95%:  91529.90
#> 
mean(B)
#> $ByOrigin
#>       Mean IBNR
#> 1981     0.0000
#> 1982   178.5882
#> 1983   694.4367
#> 1984  1768.2187
#> 1985  2875.5878
#> 1986  3759.1547
#> 1987  5620.2341
#> 1988 11217.0691
#> 1989 11432.7087
#> 1990 18004.9719
#> 
#> $Totals
#>               Total
#> Mean IBNR: 55550.97
#> 
quantile(B, c(0.75,0.95,0.99, 0.995))
#> $ByOrigin
#>        IBNR 75%  IBNR 95%  IBNR 99% IBNR 99.5%
#> 1981     0.0000     0.000     0.000      0.000
#> 1982   184.2002  1456.344  3137.262   3839.259
#> 1983  1184.7417  3028.437  5163.274   6499.323
#> 1984  2744.4657  5260.209  7923.800   9730.029
#> 1985  4137.0464  6977.161 10112.359  10743.726
#> 1986  5156.7745  8407.995 10960.905  12300.130
#> 1987  7324.1357 11375.187 14932.762  15215.631
#> 1988 14164.8761 20551.177 25404.471  26280.425
#> 1989 15208.0374 23034.658 28639.903  30241.703
#> 1990 25144.7684 45410.332 62966.352  66235.173
#> 
#> $Totals
#>               Totals
#> IBNR 75%:    66993.5
#> IBNR 95%:    91529.9
#> IBNR 99%:   109660.3
#> IBNR 99.5%: 114454.9
#>