Methods for BootChainLadder objects

summary, print, mean, and quantile methods for BootChainLadder objects

Usage

# S3 method for BootChainLadder
summary(object, probs=c(0.75,0.95), ...)

# S3 method for BootChainLadder
print(x, probs=c(0.75,0.95), ...)

# S3 method for BootChainLadder
quantile(x, probs=c(0.75, 0.95), na.rm = FALSE,
              names = TRUE, type = 7,...)

# S3 method for BootChainLadder
mean(x, ...)

# S3 method for BootChainLadder
residuals(object, ...)

Arguments

x, object

output from BootChainLadder

probs

numeric vector of probabilities with values in [0,1], see quantile for more help

na.rm

logical; if true, any NA and NaN's are removed from 'x' before the quantiles are computed, see quantile for more help

names

logical; if true, the result has a names attribute. Set to FALSE for speedup with many 'probs', see quantile for more help

type

an integer between 1 and 9 selecting one of the nine quantile algorithms detailed below to be used, see quantile

...

further arguments passed to or from other methods

Details

print.BootChainLadder calls summary.BootChainLadder and prints a formatted version of the summary. residuals.BootChainLadder gives the residual triangle of the expected chain-ladder minus the actual triangle back.

Value

summary.BootChainLadder, mean.BootChainLadder, and

quantile.BootChainLadder, give a list with two elements back:

ByOrigin

data frame with summary/mean/quantile statistics by origin period

Totals

data frame with total summary/mean/quantile statistics for all origin period

Author

Markus Gesmann

See also

Examples

B <- BootChainLadder(RAA, R=999, process.distr="gamma")
B
#> BootChainLadder(Triangle = RAA, R = 999, process.distr = "gamma")
#> 
#>      Latest Mean Ultimate Mean IBNR IBNR.S.E IBNR 75% IBNR 95%
#> 1981 18,834        18,834         0        0        0        0
#> 1982 16,704        16,907       203      731      231    1,550
#> 1983 23,466        24,109       643    1,336    1,114    3,042
#> 1984 27,067        28,767     1,700    1,942    2,651    5,138
#> 1985 26,180        28,961     2,781    2,329    3,970    7,089
#> 1986 15,852        19,529     3,677    2,518    5,109    8,726
#> 1987 12,314        17,806     5,492    3,081    7,191   11,138
#> 1988 13,112        24,259    11,147    4,902   13,896   20,112
#> 1989  5,395        16,298    10,903    5,827   14,748   20,951
#> 1990  2,063        19,045    16,982   13,832   25,282   41,978
#> 
#>                  Totals
#> Latest:         160,987
#> Mean Ultimate:  214,515
#> Mean IBNR:       53,528
#> IBNR.S.E         18,723
#> Total IBNR 75%:  65,241
#> Total IBNR 95%:  88,483
summary(B)
#> $ByOrigin
#>      Latest Mean Ultimate  Mean IBNR    SD IBNR  IBNR 75%  IBNR 95%
#> 1981  18834      18834.00     0.0000     0.0000     0.000     0.000
#> 1982  16704      16907.41   203.4084   730.5777   230.933  1550.235
#> 1983  23466      24109.15   643.1497  1335.8913  1114.337  3041.908
#> 1984  27067      28766.77  1699.7661  1942.2913  2651.436  5137.519
#> 1985  26180      28961.21  2781.2052  2329.4026  3970.310  7089.030
#> 1986  15852      19529.44  3677.4358  2517.9350  5108.808  8725.796
#> 1987  12314      17805.95  5491.9478  3080.8585  7191.174 11138.201
#> 1988  13112      24258.65 11146.6526  4901.7807 13895.756 20112.218
#> 1989   5395      16297.55 10902.5460  5827.4635 14747.970 20951.398
#> 1990   2063      19045.07 16982.0724 13831.8022 25282.071 41977.764
#> 
#> $Totals
#>                    Totals
#> Latest:         160987.00
#> Mean Ultimate:  214515.18
#> Mean IBNR:       53528.18
#> SD IBNR:         18722.98
#> Total IBNR 75%:  65240.79
#> Total IBNR 95%:  88483.43
#> 
mean(B)
#> $ByOrigin
#>       Mean IBNR
#> 1981     0.0000
#> 1982   203.4084
#> 1983   643.1497
#> 1984  1699.7661
#> 1985  2781.2052
#> 1986  3677.4358
#> 1987  5491.9478
#> 1988 11146.6526
#> 1989 10902.5460
#> 1990 16982.0724
#> 
#> $Totals
#>               Total
#> Mean IBNR: 53528.18
#> 
quantile(B, c(0.75,0.95,0.99, 0.995))
#> $ByOrigin
#>       IBNR 75%  IBNR 95%  IBNR 99% IBNR 99.5%
#> 1981     0.000     0.000     0.000      0.000
#> 1982   230.933  1550.235  2900.538   3987.729
#> 1983  1114.337  3041.908  4827.653   5646.144
#> 1984  2651.436  5137.519  8648.240   9009.846
#> 1985  3970.310  7089.030 10221.395  10640.767
#> 1986  5108.808  8725.796 11112.602  12787.805
#> 1987  7191.174 11138.201 14454.148  15148.074
#> 1988 13895.756 20112.218 25807.409  26604.713
#> 1989 14747.970 20951.398 26332.881  27648.416
#> 1990 25282.071 41977.764 60221.561  63736.572
#> 
#> $Totals
#>                Totals
#> IBNR 75%:    65240.79
#> IBNR 95%:    88483.43
#> IBNR 99%:   103582.82
#> IBNR 99.5%: 110895.75
#>