Methods for BootChainLadder objects

summary, print, mean, and quantile methods for BootChainLadder objects

Usage

# S3 method for BootChainLadder
summary(object, probs=c(0.75,0.95), ...)

# S3 method for BootChainLadder
print(x, probs=c(0.75,0.95), ...)

# S3 method for BootChainLadder
quantile(x, probs=c(0.75, 0.95), na.rm = FALSE,
              names = TRUE, type = 7,...)

# S3 method for BootChainLadder
mean(x, ...)

# S3 method for BootChainLadder
residuals(object, ...)

Arguments

x, object

output from BootChainLadder

probs

numeric vector of probabilities with values in [0,1], see quantile for more help

na.rm

logical; if true, any NA and NaN's are removed from 'x' before the quantiles are computed, see quantile for more help

names

logical; if true, the result has a names attribute. Set to FALSE for speedup with many 'probs', see quantile for more help

type

an integer between 1 and 9 selecting one of the nine quantile algorithms detailed below to be used, see quantile

...

further arguments passed to or from other methods

Details

print.BootChainLadder calls summary.BootChainLadder and prints a formatted version of the summary. residuals.BootChainLadder gives the residual triangle of the expected chain-ladder minus the actual triangle back.

Value

summary.BootChainLadder, mean.BootChainLadder, and

quantile.BootChainLadder, give a list with two elements back:

ByOrigin

data frame with summary/mean/quantile statistics by origin period

Totals

data frame with total summary/mean/quantile statistics for all origin period

Author

Markus Gesmann

See also

Examples

B <- BootChainLadder(RAA, R=999, process.distr="gamma")
B
#> BootChainLadder(Triangle = RAA, R = 999, process.distr = "gamma")
#> 
#>      Latest Mean Ultimate Mean IBNR IBNR.S.E IBNR 75% IBNR 95%
#> 1981 18,834        18,834         0        0        0        0
#> 1982 16,704        16,850       146      692      150    1,382
#> 1983 23,466        24,087       621    1,194    1,168    2,754
#> 1984 27,067        28,806     1,739    1,822    2,662    5,069
#> 1985 26,180        29,086     2,906    2,367    4,101    7,080
#> 1986 15,852        19,760     3,908    2,599    5,290    8,783
#> 1987 12,314        17,835     5,521    3,106    7,326   11,187
#> 1988 13,112        24,548    11,436    5,310   14,309   21,286
#> 1989  5,395        16,305    10,910    5,836   14,589   21,731
#> 1990  2,063        19,254    17,191   13,801   25,827   42,425
#> 
#>                  Totals
#> Latest:         160,987
#> Mean Ultimate:  215,365
#> Mean IBNR:       54,378
#> IBNR.S.E         18,640
#> Total IBNR 75%:  66,011
#> Total IBNR 95%:  86,282
summary(B)
#> $ByOrigin
#>      Latest Mean Ultimate  Mean IBNR    SD IBNR   IBNR 75%  IBNR 95%
#> 1981  18834      18834.00     0.0000     0.0000     0.0000     0.000
#> 1982  16704      16850.05   146.0486   691.8783   150.1444  1382.001
#> 1983  23466      24087.01   621.0118  1193.7651  1168.2896  2753.657
#> 1984  27067      28805.52  1738.5218  1822.3987  2661.6411  5068.927
#> 1985  26180      29085.66  2905.6642  2366.6648  4101.3538  7079.580
#> 1986  15852      19760.48  3908.4760  2598.6241  5289.6319  8783.145
#> 1987  12314      17835.44  5521.4393  3105.6157  7325.5155 11187.387
#> 1988  13112      24548.31 11436.3080  5310.4103 14308.8701 21286.398
#> 1989   5395      16304.70 10909.7036  5835.7024 14589.2538 21731.103
#> 1990   2063      19253.55 17190.5467 13800.6382 25827.4983 42425.422
#> 
#> $Totals
#>                    Totals
#> Latest:         160987.00
#> Mean Ultimate:  215364.72
#> Mean IBNR:       54377.72
#> SD IBNR:         18640.09
#> Total IBNR 75%:  66011.22
#> Total IBNR 95%:  86282.38
#> 
mean(B)
#> $ByOrigin
#>       Mean IBNR
#> 1981     0.0000
#> 1982   146.0486
#> 1983   621.0118
#> 1984  1738.5218
#> 1985  2905.6642
#> 1986  3908.4760
#> 1987  5521.4393
#> 1988 11436.3080
#> 1989 10909.7036
#> 1990 17190.5467
#> 
#> $Totals
#>               Total
#> Mean IBNR: 54377.72
#> 
quantile(B, c(0.75,0.95,0.99, 0.995))
#> $ByOrigin
#>        IBNR 75%  IBNR 95%  IBNR 99% IBNR 99.5%
#> 1981     0.0000     0.000     0.000      0.000
#> 1982   150.1444  1382.001  2681.665   3524.177
#> 1983  1168.2896  2753.657  4319.390   4694.000
#> 1984  2661.6411  5068.927  7376.296   8176.353
#> 1985  4101.3538  7079.580  9831.674  11681.464
#> 1986  5289.6319  8783.145 12140.671  13148.216
#> 1987  7325.5155 11187.387 15406.081  16624.410
#> 1988 14308.8701 21286.398 27257.820  28874.583
#> 1989 14589.2538 21731.103 26235.050  28336.410
#> 1990 25827.4983 42425.422 57812.692  59198.956
#> 
#> $Totals
#>                Totals
#> IBNR 75%:    66011.22
#> IBNR 95%:    86282.38
#> IBNR 99%:   105469.71
#> IBNR 99.5%: 109835.66
#>