Methods for BootChainLadder objects

summary, print, mean, and quantile methods for BootChainLadder objects

Usage

# S3 method for BootChainLadder
summary(object, probs=c(0.75,0.95), ...)

# S3 method for BootChainLadder
print(x, probs=c(0.75,0.95), ...)

# S3 method for BootChainLadder
quantile(x, probs=c(0.75, 0.95), na.rm = FALSE,
              names = TRUE, type = 7,...)

# S3 method for BootChainLadder
mean(x, ...)

# S3 method for BootChainLadder
residuals(object, ...)

Arguments

x, object

output from BootChainLadder

probs

numeric vector of probabilities with values in [0,1], see quantile for more help

na.rm

logical; if true, any NA and NaN's are removed from 'x' before the quantiles are computed, see quantile for more help

names

logical; if true, the result has a names attribute. Set to FALSE for speedup with many 'probs', see quantile for more help

type

an integer between 1 and 9 selecting one of the nine quantile algorithms detailed below to be used, see quantile

...

further arguments passed to or from other methods

Details

print.BootChainLadder calls summary.BootChainLadder and prints a formatted version of the summary. residuals.BootChainLadder gives the residual triangle of the expected chain-ladder minus the actual triangle back.

Value

summary.BootChainLadder, mean.BootChainLadder, and

quantile.BootChainLadder, give a list with two elements back:

ByOrigin

data frame with summary/mean/quantile statistics by origin period

Totals

data frame with total summary/mean/quantile statistics for all origin period

Author

Markus Gesmann

See also

Examples

B <- BootChainLadder(RAA, R=999, process.distr="gamma")
B
#> BootChainLadder(Triangle = RAA, R = 999, process.distr = "gamma")
#> 
#>      Latest Mean Ultimate Mean IBNR IBNR.S.E IBNR 75% IBNR 95%
#> 1981 18,834        18,834         0        0        0        0
#> 1982 16,704        16,889       185      667      228    1,437
#> 1983 23,466        24,109       643    1,284    1,151    3,099
#> 1984 27,067        28,765     1,698    1,907    2,617    5,369
#> 1985 26,180        29,046     2,866    2,362    4,089    7,199
#> 1986 15,852        19,641     3,789    2,449    5,177    8,236
#> 1987 12,314        17,833     5,519    3,179    7,394   11,435
#> 1988 13,112        24,418    11,306    5,264   14,164   20,496
#> 1989  5,395        16,581    11,186    6,118   14,924   22,574
#> 1990  2,063        20,251    18,188   14,511   25,870   44,659
#> 
#>                  Totals
#> Latest:         160,987
#> Mean Ultimate:  216,369
#> Mean IBNR:       55,382
#> IBNR.S.E         19,759
#> Total IBNR 75%:  67,160
#> Total IBNR 95%:  90,944
summary(B)
#> $ByOrigin
#>      Latest Mean Ultimate  Mean IBNR    SD IBNR   IBNR 75%  IBNR 95%
#> 1981  18834      18834.00     0.0000     0.0000     0.0000     0.000
#> 1982  16704      16889.22   185.2239   667.4091   227.9491  1437.315
#> 1983  23466      24109.14   643.1401  1283.6413  1151.4319  3098.997
#> 1984  27067      28764.96  1697.9639  1907.1005  2616.9859  5369.370
#> 1985  26180      29046.23  2866.2344  2361.7776  4088.7890  7199.128
#> 1986  15852      19641.26  3789.2560  2448.5135  5176.6712  8236.180
#> 1987  12314      17833.33  5519.3273  3178.9182  7394.1911 11434.719
#> 1988  13112      24418.39 11306.3946  5263.8864 14164.1966 20496.290
#> 1989   5395      16581.39 11186.3942  6118.0696 14924.0297 22574.229
#> 1990   2063      20250.75 18187.7477 14511.0333 25870.3570 44658.786
#> 
#> $Totals
#>                    Totals
#> Latest:         160987.00
#> Mean Ultimate:  216368.68
#> Mean IBNR:       55381.68
#> SD IBNR:         19759.12
#> Total IBNR 75%:  67160.33
#> Total IBNR 95%:  90943.78
#> 
mean(B)
#> $ByOrigin
#>       Mean IBNR
#> 1981     0.0000
#> 1982   185.2239
#> 1983   643.1401
#> 1984  1697.9639
#> 1985  2866.2344
#> 1986  3789.2560
#> 1987  5519.3273
#> 1988 11306.3946
#> 1989 11186.3942
#> 1990 18187.7477
#> 
#> $Totals
#>               Total
#> Mean IBNR: 55381.68
#> 
quantile(B, c(0.75,0.95,0.99, 0.995))
#> $ByOrigin
#>        IBNR 75%  IBNR 95%  IBNR 99% IBNR 99.5%
#> 1981     0.0000     0.000     0.000      0.000
#> 1982   227.9491  1437.315  2546.777   3463.506
#> 1983  1151.4319  3098.997  4841.886   5289.209
#> 1984  2616.9859  5369.370  7119.752   8446.643
#> 1985  4088.7890  7199.128 10058.987  10951.511
#> 1986  5176.6712  8236.180 11019.671  12332.035
#> 1987  7394.1911 11434.719 14644.543  16233.936
#> 1988 14164.1966 20496.290 26391.391  31404.260
#> 1989 14924.0297 22574.229 28260.166  29820.926
#> 1990 25870.3570 44658.786 58287.744  67238.037
#> 
#> $Totals
#>                Totals
#> IBNR 75%:    67160.33
#> IBNR 95%:    90943.78
#> IBNR 99%:   108418.11
#> IBNR 99.5%: 117565.11
#>