Methods for BootChainLadder objects

summary, print, mean, and quantile methods for BootChainLadder objects

Usage

# S3 method for BootChainLadder
summary(object, probs=c(0.75,0.95), ...)

# S3 method for BootChainLadder
print(x, probs=c(0.75,0.95), ...)

# S3 method for BootChainLadder
quantile(x, probs=c(0.75, 0.95), na.rm = FALSE,
              names = TRUE, type = 7,...)

# S3 method for BootChainLadder
mean(x, ...)

# S3 method for BootChainLadder
residuals(object, ...)

Arguments

x, object

output from BootChainLadder

probs

numeric vector of probabilities with values in [0,1], see quantile for more help

na.rm

logical; if true, any NA and NaN's are removed from 'x' before the quantiles are computed, see quantile for more help

names

logical; if true, the result has a names attribute. Set to FALSE for speedup with many 'probs', see quantile for more help

type

an integer between 1 and 9 selecting one of the nine quantile algorithms detailed below to be used, see quantile

...

further arguments passed to or from other methods

Details

print.BootChainLadder calls summary.BootChainLadder and prints a formatted version of the summary. residuals.BootChainLadder gives the residual triangle of the expected chain-ladder minus the actual triangle back.

Value

summary.BootChainLadder, mean.BootChainLadder, and

quantile.BootChainLadder, give a list with two elements back:

ByOrigin

data frame with summary/mean/quantile statistics by origin period

Totals

data frame with total summary/mean/quantile statistics for all origin period

Author

Markus Gesmann

See also

Examples

B <- BootChainLadder(RAA, R=999, process.distr="gamma")
B
#> BootChainLadder(Triangle = RAA, R = 999, process.distr = "gamma")
#> 
#>      Latest Mean Ultimate Mean IBNR IBNR.S.E IBNR 75% IBNR 95%
#> 1981 18,834        18,834         0        0        0        0
#> 1982 16,704        16,866       162      657      172    1,323
#> 1983 23,466        24,150       684    1,245    1,215    3,154
#> 1984 27,067        28,866     1,799    1,909    2,733    5,216
#> 1985 26,180        29,108     2,928    2,345    4,256    7,140
#> 1986 15,852        19,735     3,883    2,603    5,143    9,026
#> 1987 12,314        17,663     5,349    2,995    7,132   11,008
#> 1988 13,112        24,188    11,076    5,178   14,260   20,162
#> 1989  5,395        16,022    10,627    5,855   14,074   21,398
#> 1990  2,063        20,219    18,156   14,395   25,201   47,290
#> 
#>                  Totals
#> Latest:         160,987
#> Mean Ultimate:  215,652
#> Mean IBNR:       54,665
#> IBNR.S.E         19,624
#> Total IBNR 75%:  65,020
#> Total IBNR 95%:  92,462
summary(B)
#> $ByOrigin
#>      Latest Mean Ultimate  Mean IBNR    SD IBNR   IBNR 75%  IBNR 95%
#> 1981  18834      18834.00     0.0000     0.0000     0.0000     0.000
#> 1982  16704      16866.47   162.4673   657.1584   171.7375  1323.439
#> 1983  23466      24150.46   684.4614  1244.9706  1214.5511  3154.463
#> 1984  27067      28866.19  1799.1876  1908.5713  2733.0506  5216.197
#> 1985  26180      29107.93  2927.9268  2345.1397  4255.7233  7139.699
#> 1986  15852      19735.24  3883.2449  2602.5710  5142.9212  9025.907
#> 1987  12314      17663.13  5349.1275  2994.5675  7131.7478 11007.559
#> 1988  13112      24187.60 11075.5965  5178.3273 14260.3171 20161.817
#> 1989   5395      16021.68 10626.6769  5855.4171 14074.1888 21397.764
#> 1990   2063      20219.25 18156.2495 14395.2559 25201.2344 47289.738
#> 
#> $Totals
#>                    Totals
#> Latest:         160987.00
#> Mean Ultimate:  215651.94
#> Mean IBNR:       54664.94
#> SD IBNR:         19623.66
#> Total IBNR 75%:  65020.34
#> Total IBNR 95%:  92461.57
#> 
mean(B)
#> $ByOrigin
#>       Mean IBNR
#> 1981     0.0000
#> 1982   162.4673
#> 1983   684.4614
#> 1984  1799.1876
#> 1985  2927.9268
#> 1986  3883.2449
#> 1987  5349.1275
#> 1988 11075.5965
#> 1989 10626.6769
#> 1990 18156.2495
#> 
#> $Totals
#>               Total
#> Mean IBNR: 54664.94
#> 
quantile(B, c(0.75,0.95,0.99, 0.995))
#> $ByOrigin
#>        IBNR 75%  IBNR 95%  IBNR 99% IBNR 99.5%
#> 1981     0.0000     0.000     0.000      0.000
#> 1982   171.7375  1323.439  2511.204   3170.612
#> 1983  1214.5511  3154.463  4557.781   4777.060
#> 1984  2733.0506  5216.197  7632.916   8359.332
#> 1985  4255.7233  7139.699 10072.729  10754.285
#> 1986  5142.9212  9025.907 12175.047  13817.119
#> 1987  7131.7478 11007.559 14025.159  14800.978
#> 1988 14260.3171 20161.817 27048.892  29685.633
#> 1989 14074.1888 21397.764 26512.657  30904.845
#> 1990 25201.2344 47289.738 60929.822  68012.796
#> 
#> $Totals
#>                Totals
#> IBNR 75%:    65020.34
#> IBNR 95%:    92461.57
#> IBNR 99%:   110025.55
#> IBNR 99.5%: 119198.02
#>