Methods for BootChainLadder objects

summary, print, mean, and quantile methods for BootChainLadder objects

Usage

# S3 method for BootChainLadder
summary(object, probs=c(0.75,0.95), ...)

# S3 method for BootChainLadder
print(x, probs=c(0.75,0.95), ...)

# S3 method for BootChainLadder
quantile(x, probs=c(0.75, 0.95), na.rm = FALSE,
              names = TRUE, type = 7,...)

# S3 method for BootChainLadder
mean(x, ...)

# S3 method for BootChainLadder
residuals(object, ...)

Arguments

x, object

output from BootChainLadder

probs

numeric vector of probabilities with values in [0,1], see quantile for more help

na.rm

logical; if true, any NA and NaN's are removed from 'x' before the quantiles are computed, see quantile for more help

names

logical; if true, the result has a names attribute. Set to FALSE for speedup with many 'probs', see quantile for more help

type

an integer between 1 and 9 selecting one of the nine quantile algorithms detailed below to be used, see quantile

...

further arguments passed to or from other methods

Details

print.BootChainLadder calls summary.BootChainLadder and prints a formatted version of the summary. residuals.BootChainLadder gives the residual triangle of the expected chain-ladder minus the actual triangle back.

Value

summary.BootChainLadder, mean.BootChainLadder, and

quantile.BootChainLadder, give a list with two elements back:

ByOrigin

data frame with summary/mean/quantile statistics by origin period

Totals

data frame with total summary/mean/quantile statistics for all origin period

Author

Markus Gesmann

See also

Examples

B <- BootChainLadder(RAA, R=999, process.distr="gamma")
B
#> BootChainLadder(Triangle = RAA, R = 999, process.distr = "gamma")
#> 
#>      Latest Mean Ultimate Mean IBNR IBNR.S.E IBNR 75% IBNR 95%
#> 1981 18,834        18,834         0        0        0        0
#> 1982 16,704        16,856       152      695      195    1,269
#> 1983 23,466        24,125       659    1,243    1,053    3,178
#> 1984 27,067        28,651     1,584    1,800    2,602    5,006
#> 1985 26,180        28,998     2,818    2,446    4,254    7,306
#> 1986 15,852        19,597     3,745    2,431    5,116    8,208
#> 1987 12,314        17,919     5,605    3,195    7,481   11,533
#> 1988 13,112        24,341    11,229    5,172   14,116   20,828
#> 1989  5,395        16,588    11,193    6,221   14,727   22,494
#> 1990  2,063        19,489    17,426   13,384   24,792   43,403
#> 
#>                  Totals
#> Latest:         160,987
#> Mean Ultimate:  215,398
#> Mean IBNR:       54,411
#> IBNR.S.E         18,809
#> Total IBNR 75%:  66,272
#> Total IBNR 95%:  87,680
summary(B)
#> $ByOrigin
#>      Latest Mean Ultimate  Mean IBNR    SD IBNR   IBNR 75%  IBNR 95%
#> 1981  18834      18834.00     0.0000     0.0000     0.0000     0.000
#> 1982  16704      16855.68   151.6837   694.5177   195.0983  1269.251
#> 1983  23466      24125.12   659.1235  1243.4357  1053.2866  3177.792
#> 1984  27067      28650.78  1583.7759  1799.9282  2601.5153  5006.049
#> 1985  26180      28997.67  2817.6735  2446.2085  4254.4665  7305.904
#> 1986  15852      19597.14  3745.1389  2430.7951  5116.4696  8207.876
#> 1987  12314      17918.95  5604.9486  3194.8593  7480.9498 11533.372
#> 1988  13112      24341.15 11229.1488  5171.7119 14115.5300 20827.961
#> 1989   5395      16588.18 11193.1753  6221.4982 14726.7310 22494.463
#> 1990   2063      19489.22 17426.2207 13384.3809 24791.6874 43403.387
#> 
#> $Totals
#>                    Totals
#> Latest:         160987.00
#> Mean Ultimate:  215397.89
#> Mean IBNR:       54410.89
#> SD IBNR:         18808.97
#> Total IBNR 75%:  66272.15
#> Total IBNR 95%:  87679.58
#> 
mean(B)
#> $ByOrigin
#>       Mean IBNR
#> 1981     0.0000
#> 1982   151.6837
#> 1983   659.1235
#> 1984  1583.7759
#> 1985  2817.6735
#> 1986  3745.1389
#> 1987  5604.9486
#> 1988 11229.1488
#> 1989 11193.1753
#> 1990 17426.2207
#> 
#> $Totals
#>               Total
#> Mean IBNR: 54410.89
#> 
quantile(B, c(0.75,0.95,0.99, 0.995))
#> $ByOrigin
#>        IBNR 75%  IBNR 95%  IBNR 99% IBNR 99.5%
#> 1981     0.0000     0.000     0.000      0.000
#> 1982   195.0983  1269.251  2728.370   3348.545
#> 1983  1053.2866  3177.792  5210.840   5623.010
#> 1984  2601.5153  5006.049  7200.219   7508.235
#> 1985  4254.4665  7305.904 10018.876  10906.958
#> 1986  5116.4696  8207.876 11552.824  12538.014
#> 1987  7480.9498 11533.372 15248.452  16536.096
#> 1988 14115.5300 20827.961 27156.332  27973.752
#> 1989 14726.7310 22494.463 29331.673  30387.541
#> 1990 24791.6874 43403.387 55867.475  61597.066
#> 
#> $Totals
#>                Totals
#> IBNR 75%:    66272.15
#> IBNR 95%:    87679.58
#> IBNR 99%:   105267.08
#> IBNR 99.5%: 111434.74
#>