Methods for BootChainLadder objects

summary, print, mean, and quantile methods for BootChainLadder objects

Usage

# S3 method for BootChainLadder
summary(object, probs=c(0.75,0.95), ...)

# S3 method for BootChainLadder
print(x, probs=c(0.75,0.95), ...)

# S3 method for BootChainLadder
quantile(x, probs=c(0.75, 0.95), na.rm = FALSE,
              names = TRUE, type = 7,...)

# S3 method for BootChainLadder
mean(x, ...)

# S3 method for BootChainLadder
residuals(object, ...)

Arguments

x, object

output from BootChainLadder

probs

numeric vector of probabilities with values in [0,1], see quantile for more help

na.rm

logical; if true, any NA and NaN's are removed from 'x' before the quantiles are computed, see quantile for more help

names

logical; if true, the result has a names attribute. Set to FALSE for speedup with many 'probs', see quantile for more help

type

an integer between 1 and 9 selecting one of the nine quantile algorithms detailed below to be used, see quantile

...

further arguments passed to or from other methods

Details

print.BootChainLadder calls summary.BootChainLadder and prints a formatted version of the summary. residuals.BootChainLadder gives the residual triangle of the expected chain-ladder minus the actual triangle back.

Value

summary.BootChainLadder, mean.BootChainLadder, and

quantile.BootChainLadder, give a list with two elements back:

ByOrigin

data frame with summary/mean/quantile statistics by origin period

Totals

data frame with total summary/mean/quantile statistics for all origin period

Author

Markus Gesmann

See also

Examples

B <- BootChainLadder(RAA, R=999, process.distr="gamma")
B
#> BootChainLadder(Triangle = RAA, R = 999, process.distr = "gamma")
#> 
#>      Latest Mean Ultimate Mean IBNR IBNR.S.E IBNR 75% IBNR 95%
#> 1981 18,834        18,834         0        0        0        0
#> 1982 16,704        16,857       153      615      175    1,399
#> 1983 23,466        24,156       690    1,290    1,168    3,061
#> 1984 27,067        28,774     1,707    1,912    2,716    5,327
#> 1985 26,180        29,113     2,933    2,344    4,034    7,540
#> 1986 15,852        19,558     3,706    2,410    5,049    8,198
#> 1987 12,314        17,924     5,610    3,109    7,274   11,327
#> 1988 13,112        24,082    10,970    4,793   13,646   19,854
#> 1989  5,395        16,248    10,853    5,912   14,729   21,772
#> 1990  2,063        19,965    17,902   14,155   25,756   44,590
#> 
#>                  Totals
#> Latest:         160,987
#> Mean Ultimate:  215,511
#> Mean IBNR:       54,524
#> IBNR.S.E         19,151
#> Total IBNR 75%:  65,141
#> Total IBNR 95%:  89,400
summary(B)
#> $ByOrigin
#>      Latest Mean Ultimate  Mean IBNR    SD IBNR   IBNR 75%  IBNR 95%
#> 1981  18834      18834.00     0.0000     0.0000     0.0000     0.000
#> 1982  16704      16857.00   153.0049   614.8147   175.4691  1398.748
#> 1983  23466      24155.96   689.9622  1290.1228  1168.2250  3060.536
#> 1984  27067      28773.79  1706.7911  1911.8688  2715.8793  5327.279
#> 1985  26180      29113.13  2933.1299  2344.1692  4034.4490  7540.490
#> 1986  15852      19557.96  3705.9607  2410.3497  5049.1654  8197.600
#> 1987  12314      17924.12  5610.1218  3108.7216  7273.7165 11327.279
#> 1988  13112      24081.83 10969.8275  4792.8530 13645.5460 19854.487
#> 1989   5395      16247.79 10852.7947  5911.6769 14728.7955 21772.000
#> 1990   2063      19965.31 17902.3057 14155.0624 25756.2306 44589.546
#> 
#> $Totals
#>                    Totals
#> Latest:         160987.00
#> Mean Ultimate:  215510.90
#> Mean IBNR:       54523.90
#> SD IBNR:         19151.46
#> Total IBNR 75%:  65140.85
#> Total IBNR 95%:  89399.81
#> 
mean(B)
#> $ByOrigin
#>       Mean IBNR
#> 1981     0.0000
#> 1982   153.0049
#> 1983   689.9622
#> 1984  1706.7911
#> 1985  2933.1299
#> 1986  3705.9607
#> 1987  5610.1218
#> 1988 10969.8275
#> 1989 10852.7947
#> 1990 17902.3057
#> 
#> $Totals
#>              Total
#> Mean IBNR: 54523.9
#> 
quantile(B, c(0.75,0.95,0.99, 0.995))
#> $ByOrigin
#>        IBNR 75%  IBNR 95%  IBNR 99% IBNR 99.5%
#> 1981     0.0000     0.000     0.000      0.000
#> 1982   175.4691  1398.748  2427.969   2909.365
#> 1983  1168.2250  3060.536  5257.348   5760.693
#> 1984  2715.8793  5327.279  8054.904   9608.577
#> 1985  4034.4490  7540.490  9717.630  11536.719
#> 1986  5049.1654  8197.600 10302.805  11426.948
#> 1987  7273.7165 11327.279 15190.056  17439.368
#> 1988 13645.5460 19854.487 24155.379  26220.426
#> 1989 14728.7955 21772.000 26607.402  28994.461
#> 1990 25756.2306 44589.546 60508.090  64907.384
#> 
#> $Totals
#>                Totals
#> IBNR 75%:    65140.85
#> IBNR 95%:    89399.81
#> IBNR 99%:   111452.89
#> IBNR 99.5%: 120305.82
#>