Methods for BootChainLadder objects

summary, print, mean, and quantile methods for BootChainLadder objects

Usage

# S3 method for BootChainLadder
summary(object, probs=c(0.75,0.95), ...)

# S3 method for BootChainLadder
print(x, probs=c(0.75,0.95), ...)

# S3 method for BootChainLadder
quantile(x, probs=c(0.75, 0.95), na.rm = FALSE,
              names = TRUE, type = 7,...)

# S3 method for BootChainLadder
mean(x, ...)

# S3 method for BootChainLadder
residuals(object, ...)

Arguments

x, object

output from BootChainLadder

probs

numeric vector of probabilities with values in [0,1], see quantile for more help

na.rm

logical; if true, any NA and NaN's are removed from 'x' before the quantiles are computed, see quantile for more help

names

logical; if true, the result has a names attribute. Set to FALSE for speedup with many 'probs', see quantile for more help

type

an integer between 1 and 9 selecting one of the nine quantile algorithms detailed below to be used, see quantile

...

further arguments passed to or from other methods

Details

print.BootChainLadder calls summary.BootChainLadder and prints a formatted version of the summary. residuals.BootChainLadder gives the residual triangle of the expected chain-ladder minus the actual triangle back.

Value

summary.BootChainLadder, mean.BootChainLadder, and

quantile.BootChainLadder, give a list with two elements back:

ByOrigin

data frame with summary/mean/quantile statistics by origin period

Totals

data frame with total summary/mean/quantile statistics for all origin period

Author

Markus Gesmann

See also

Examples

B <- BootChainLadder(RAA, R=999, process.distr="gamma")
B
#> BootChainLadder(Triangle = RAA, R = 999, process.distr = "gamma")
#> 
#>      Latest Mean Ultimate Mean IBNR IBNR.S.E IBNR 75% IBNR 95%
#> 1981 18,834        18,834         0        0        0        0
#> 1982 16,704        16,865       161      737      166    1,333
#> 1983 23,466        24,058       592    1,207    1,012    2,919
#> 1984 27,067        28,759     1,692    1,918    2,673    5,198
#> 1985 26,180        28,875     2,695    2,351    3,955    7,068
#> 1986 15,852        19,522     3,670    2,453    5,140    8,381
#> 1987 12,314        17,838     5,524    3,136    7,453   11,095
#> 1988 13,112        24,219    11,107    4,974   14,138   19,969
#> 1989  5,395        16,136    10,741    5,938   14,391   21,233
#> 1990  2,063        19,136    17,073   14,043   25,301   43,247
#> 
#>                  Totals
#> Latest:         160,987
#> Mean Ultimate:  214,242
#> Mean IBNR:       53,255
#> IBNR.S.E         18,710
#> Total IBNR 75%:  64,953
#> Total IBNR 95%:  86,908
summary(B)
#> $ByOrigin
#>      Latest Mean Ultimate  Mean IBNR    SD IBNR   IBNR 75%  IBNR 95%
#> 1981  18834      18834.00     0.0000     0.0000     0.0000     0.000
#> 1982  16704      16864.95   160.9522   736.9384   166.1873  1332.573
#> 1983  23466      24057.91   591.9100  1206.7523  1011.6133  2919.227
#> 1984  27067      28758.85  1691.8529  1918.4618  2673.3452  5197.600
#> 1985  26180      28875.31  2695.3132  2351.0984  3955.0866  7067.527
#> 1986  15852      19522.32  3670.3196  2452.6660  5140.3760  8381.008
#> 1987  12314      17837.54  5523.5377  3136.3097  7452.5814 11094.907
#> 1988  13112      24219.18 11107.1825  4974.4845 14138.4489 19968.989
#> 1989   5395      16136.17 10741.1658  5938.3137 14390.5292 21232.929
#> 1990   2063      19135.88 17072.8755 14042.9609 25301.1645 43246.780
#> 
#> $Totals
#>                    Totals
#> Latest:         160987.00
#> Mean Ultimate:  214242.11
#> Mean IBNR:       53255.11
#> SD IBNR:         18710.32
#> Total IBNR 75%:  64953.25
#> Total IBNR 95%:  86907.82
#> 
mean(B)
#> $ByOrigin
#>       Mean IBNR
#> 1981     0.0000
#> 1982   160.9522
#> 1983   591.9100
#> 1984  1691.8529
#> 1985  2695.3132
#> 1986  3670.3196
#> 1987  5523.5377
#> 1988 11107.1825
#> 1989 10741.1658
#> 1990 17072.8755
#> 
#> $Totals
#>               Total
#> Mean IBNR: 53255.11
#> 
quantile(B, c(0.75,0.95,0.99, 0.995))
#> $ByOrigin
#>        IBNR 75%  IBNR 95%  IBNR 99% IBNR 99.5%
#> 1981     0.0000     0.000     0.000      0.000
#> 1982   166.1873  1332.573  3032.410   3766.306
#> 1983  1011.6133  2919.227  4949.029   5551.793
#> 1984  2673.3452  5197.600  7382.587   8778.927
#> 1985  3955.0866  7067.527  9579.339  11319.757
#> 1986  5140.3760  8381.008 10648.516  11455.209
#> 1987  7452.5814 11094.907 13770.296  15245.391
#> 1988 14138.4489 19968.989 24759.324  26387.626
#> 1989 14390.5292 21232.929 27763.661  31511.650
#> 1990 25301.1645 43246.780 60890.623  64828.623
#> 
#> $Totals
#>                Totals
#> IBNR 75%:    64953.25
#> IBNR 95%:    86907.82
#> IBNR 99%:   104282.12
#> IBNR 99.5%: 106481.03
#>