Methods for BootChainLadder objects

summary, print, mean, and quantile methods for BootChainLadder objects

Usage

# S3 method for BootChainLadder
summary(object, probs=c(0.75,0.95), ...)

# S3 method for BootChainLadder
print(x, probs=c(0.75,0.95), ...)

# S3 method for BootChainLadder
quantile(x, probs=c(0.75, 0.95), na.rm = FALSE,
              names = TRUE, type = 7,...)

# S3 method for BootChainLadder
mean(x, ...)

# S3 method for BootChainLadder
residuals(object, ...)

Arguments

x, object

output from BootChainLadder

probs

numeric vector of probabilities with values in [0,1], see quantile for more help

na.rm

logical; if true, any NA and NaN's are removed from 'x' before the quantiles are computed, see quantile for more help

names

logical; if true, the result has a names attribute. Set to FALSE for speedup with many 'probs', see quantile for more help

type

an integer between 1 and 9 selecting one of the nine quantile algorithms detailed below to be used, see quantile

...

further arguments passed to or from other methods

Details

print.BootChainLadder calls summary.BootChainLadder and prints a formatted version of the summary. residuals.BootChainLadder gives the residual triangle of the expected chain-ladder minus the actual triangle back.

Value

summary.BootChainLadder, mean.BootChainLadder, and quantile.BootChainLadder, give a list with two elements back:

ByOrigin

data frame with summary/mean/quantile statistics by origin period

Totals

data frame with total summary/mean/quantile statistics for all origin period

Author

Markus Gesmann

See also

Examples

B <- BootChainLadder(RAA, R=999, process.distr="gamma")
B
#> BootChainLadder(Triangle = RAA, R = 999, process.distr = "gamma")
#> 
#>      Latest Mean Ultimate Mean IBNR IBNR.S.E IBNR 75% IBNR 95%
#> 1981 18,834        18,834         0        0        0        0
#> 1982 16,704        16,880       176      670      191    1,403
#> 1983 23,466        24,126       660    1,329    1,074    3,185
#> 1984 27,067        28,745     1,678    1,918    2,519    5,456
#> 1985 26,180        28,996     2,816    2,242    4,054    7,094
#> 1986 15,852        19,599     3,747    2,434    5,022    8,077
#> 1987 12,314        17,791     5,477    3,282    7,243   11,903
#> 1988 13,112        24,445    11,333    5,062   13,997   21,002
#> 1989  5,395        16,528    11,133    6,113   14,652   21,967
#> 1990  2,063        19,288    17,225   13,849   24,876   41,995
#> 
#>                  Totals
#> Latest:         160,987
#> Mean Ultimate:  215,231
#> Mean IBNR:       54,244
#> IBNR.S.E         18,841
#> Total IBNR 75%:  65,275
#> Total IBNR 95%:  88,898
summary(B)
#> $ByOrigin
#>      Latest Mean Ultimate  Mean IBNR    SD IBNR   IBNR 75%  IBNR 95%
#> 1981  18834      18834.00     0.0000     0.0000     0.0000     0.000
#> 1982  16704      16879.59   175.5935   670.4753   191.3748  1402.974
#> 1983  23466      24125.89   659.8913  1329.0103  1074.1196  3184.566
#> 1984  27067      28744.86  1677.8636  1917.8316  2519.4934  5455.979
#> 1985  26180      28995.88  2815.8822  2242.0592  4053.5986  7094.064
#> 1986  15852      19599.00  3747.0015  2434.2990  5022.1064  8076.817
#> 1987  12314      17790.67  5476.6673  3282.2553  7242.5500 11903.240
#> 1988  13112      24445.38 11333.3788  5061.8672 13996.8977 21001.747
#> 1989   5395      16527.95 11132.9486  6112.9996 14652.2336 21966.718
#> 1990   2063      19287.75 17224.7524 13849.2868 24876.2658 41995.369
#> 
#> $Totals
#>                    Totals
#> Latest:         160987.00
#> Mean Ultimate:  215230.98
#> Mean IBNR:       54243.98
#> SD IBNR:         18840.84
#> Total IBNR 75%:  65275.23
#> Total IBNR 95%:  88898.17
#> 
mean(B)
#> $ByOrigin
#>       Mean IBNR
#> 1981     0.0000
#> 1982   175.5935
#> 1983   659.8913
#> 1984  1677.8636
#> 1985  2815.8822
#> 1986  3747.0015
#> 1987  5476.6673
#> 1988 11333.3788
#> 1989 11132.9486
#> 1990 17224.7524
#> 
#> $Totals
#>               Total
#> Mean IBNR: 54243.98
#> 
quantile(B, c(0.75,0.95,0.99, 0.995))
#> $ByOrigin
#>        IBNR 75%  IBNR 95%  IBNR 99% IBNR 99.5%
#> 1981     0.0000     0.000     0.000      0.000
#> 1982   191.3748  1402.974  2497.929   3447.650
#> 1983  1074.1196  3184.566  5085.415   6463.014
#> 1984  2519.4934  5455.979  7762.498   8777.323
#> 1985  4053.5986  7094.064  9332.709  10688.865
#> 1986  5022.1064  8076.817 11637.252  13733.534
#> 1987  7242.5500 11903.240 14979.340  15879.580
#> 1988 13996.8977 21001.747 26066.362  27700.895
#> 1989 14652.2336 21966.718 28508.727  30002.716
#> 1990 24876.2658 41995.369 56028.649  63272.294
#> 
#> $Totals
#>                Totals
#> IBNR 75%:    65275.23
#> IBNR 95%:    88898.17
#> IBNR 99%:   108807.89
#> IBNR 99.5%: 111519.03
#>