Methods for BootChainLadder objects

summary, print, mean, and quantile methods for BootChainLadder objects

Usage

# S3 method for BootChainLadder
summary(object, probs=c(0.75,0.95), ...)

# S3 method for BootChainLadder
print(x, probs=c(0.75,0.95), ...)

# S3 method for BootChainLadder
quantile(x, probs=c(0.75, 0.95), na.rm = FALSE,
              names = TRUE, type = 7,...)

# S3 method for BootChainLadder
mean(x, ...)

# S3 method for BootChainLadder
residuals(object, ...)

Arguments

x, object

output from BootChainLadder

probs

numeric vector of probabilities with values in [0,1], see quantile for more help

na.rm

logical; if true, any NA and NaN's are removed from 'x' before the quantiles are computed, see quantile for more help

names

logical; if true, the result has a names attribute. Set to FALSE for speedup with many 'probs', see quantile for more help

type

an integer between 1 and 9 selecting one of the nine quantile algorithms detailed below to be used, see quantile

...

further arguments passed to or from other methods

Details

print.BootChainLadder calls summary.BootChainLadder and prints a formatted version of the summary. residuals.BootChainLadder gives the residual triangle of the expected chain-ladder minus the actual triangle back.

Value

summary.BootChainLadder, mean.BootChainLadder, and

quantile.BootChainLadder, give a list with two elements back:

ByOrigin

data frame with summary/mean/quantile statistics by origin period

Totals

data frame with total summary/mean/quantile statistics for all origin period

Author

Markus Gesmann

See also

Examples

B <- BootChainLadder(RAA, R=999, process.distr="gamma")
B
#> BootChainLadder(Triangle = RAA, R = 999, process.distr = "gamma")
#> 
#>      Latest Mean Ultimate Mean IBNR IBNR.S.E IBNR 75% IBNR 95%
#> 1981 18,834        18,834         0        0        0        0
#> 1982 16,704        16,867       163      678      150    1,448
#> 1983 23,466        24,144       678    1,330    1,187    3,215
#> 1984 27,067        28,795     1,728    1,959    2,696    5,570
#> 1985 26,180        29,023     2,843    2,362    4,073    7,161
#> 1986 15,852        19,603     3,751    2,539    5,239    8,671
#> 1987 12,314        17,994     5,680    3,321    7,360   12,171
#> 1988 13,112        24,492    11,380    5,235   14,301   21,295
#> 1989  5,395        15,915    10,520    5,996   14,178   21,200
#> 1990  2,063        18,717    16,654   13,403   24,016   41,182
#> 
#>                  Totals
#> Latest:         160,987
#> Mean Ultimate:  214,384
#> Mean IBNR:       53,397
#> IBNR.S.E         18,360
#> Total IBNR 75%:  63,788
#> Total IBNR 95%:  86,447
summary(B)
#> $ByOrigin
#>      Latest Mean Ultimate  Mean IBNR    SD IBNR  IBNR 75%  IBNR 95%
#> 1981  18834      18834.00     0.0000     0.0000     0.000     0.000
#> 1982  16704      16866.52   162.5215   678.3069   150.481  1447.649
#> 1983  23466      24143.85   677.8504  1329.8081  1186.787  3215.363
#> 1984  27067      28795.39  1728.3902  1959.4768  2696.289  5570.100
#> 1985  26180      29023.47  2843.4740  2362.1993  4072.708  7160.636
#> 1986  15852      19602.93  3750.9335  2539.4142  5238.995  8671.025
#> 1987  12314      17993.90  5679.8978  3320.7037  7359.623 12171.359
#> 1988  13112      24491.83 11379.8256  5234.8685 14301.402 21295.009
#> 1989   5395      15914.71 10519.7117  5996.2348 14177.915 21199.812
#> 1990   2063      18717.47 16654.4733 13402.5006 24015.640 41182.462
#> 
#> $Totals
#>                    Totals
#> Latest:         160987.00
#> Mean Ultimate:  214384.08
#> Mean IBNR:       53397.08
#> SD IBNR:         18359.90
#> Total IBNR 75%:  63787.97
#> Total IBNR 95%:  86446.67
#> 
mean(B)
#> $ByOrigin
#>       Mean IBNR
#> 1981     0.0000
#> 1982   162.5215
#> 1983   677.8504
#> 1984  1728.3902
#> 1985  2843.4740
#> 1986  3750.9335
#> 1987  5679.8978
#> 1988 11379.8256
#> 1989 10519.7117
#> 1990 16654.4733
#> 
#> $Totals
#>               Total
#> Mean IBNR: 53397.08
#> 
quantile(B, c(0.75,0.95,0.99, 0.995))
#> $ByOrigin
#>       IBNR 75%  IBNR 95%  IBNR 99% IBNR 99.5%
#> 1981     0.000     0.000     0.000      0.000
#> 1982   150.481  1447.649  3005.498   3254.388
#> 1983  1186.787  3215.363  5080.704   5822.953
#> 1984  2696.289  5570.100  7504.075   8308.345
#> 1985  4072.708  7160.636  9846.713  10862.687
#> 1986  5238.995  8671.025 11648.986  12561.183
#> 1987  7359.623 12171.359 16244.625  17836.929
#> 1988 14301.402 21295.009 26262.953  28327.552
#> 1989 14177.915 21199.812 27290.460  29369.176
#> 1990 24015.640 41182.462 58815.368  61473.273
#> 
#> $Totals
#>                Totals
#> IBNR 75%:    63787.97
#> IBNR 95%:    86446.67
#> IBNR 99%:   102319.72
#> IBNR 99.5%: 110974.20
#>