Methods for BootChainLadder objects

summary, print, mean, and quantile methods for BootChainLadder objects

Usage

# S3 method for BootChainLadder
summary(object, probs=c(0.75,0.95), ...)

# S3 method for BootChainLadder
print(x, probs=c(0.75,0.95), ...)

# S3 method for BootChainLadder
quantile(x, probs=c(0.75, 0.95), na.rm = FALSE,
              names = TRUE, type = 7,...)

# S3 method for BootChainLadder
mean(x, ...)

# S3 method for BootChainLadder
residuals(object, ...)

Arguments

x, object

output from BootChainLadder

probs

numeric vector of probabilities with values in [0,1], see quantile for more help

na.rm

logical; if true, any NA and NaN's are removed from 'x' before the quantiles are computed, see quantile for more help

names

logical; if true, the result has a names attribute. Set to FALSE for speedup with many 'probs', see quantile for more help

type

an integer between 1 and 9 selecting one of the nine quantile algorithms detailed below to be used, see quantile

...

further arguments passed to or from other methods

Details

print.BootChainLadder calls summary.BootChainLadder and prints a formatted version of the summary. residuals.BootChainLadder gives the residual triangle of the expected chain-ladder minus the actual triangle back.

Value

summary.BootChainLadder, mean.BootChainLadder, and quantile.BootChainLadder, give a list with two elements back:

ByOrigin

data frame with summary/mean/quantile statistics by origin period

Totals

data frame with total summary/mean/quantile statistics for all origin period

Author

Markus Gesmann

See also

Examples

B <- BootChainLadder(RAA, R=999, process.distr="gamma")
B
#> BootChainLadder(Triangle = RAA, R = 999, process.distr = "gamma")
#> 
#>      Latest Mean Ultimate Mean IBNR IBNR.S.E IBNR 75% IBNR 95%
#> 1981 18,834        18,834         0        0        0        0
#> 1982 16,704        16,883       179      721      240    1,402
#> 1983 23,466        24,189       723    1,316    1,242    3,243
#> 1984 27,067        28,865     1,798    2,026    2,707    5,598
#> 1985 26,180        29,039     2,859    2,314    3,920    7,187
#> 1986 15,852        19,550     3,698    2,639    5,115    8,929
#> 1987 12,314        17,889     5,575    3,080    7,318   11,359
#> 1988 13,112        24,067    10,955    4,959   13,946   19,855
#> 1989  5,395        16,034    10,639    5,859   14,072   21,121
#> 1990  2,063        18,913    16,850   13,513   23,830   40,609
#> 
#>                  Totals
#> Latest:         160,987
#> Mean Ultimate:  214,264
#> Mean IBNR:       53,277
#> IBNR.S.E         18,348
#> Total IBNR 75%:  62,991
#> Total IBNR 95%:  86,975
summary(B)
#> $ByOrigin
#>      Latest Mean Ultimate  Mean IBNR    SD IBNR   IBNR 75%  IBNR 95%
#> 1981  18834      18834.00     0.0000     0.0000     0.0000     0.000
#> 1982  16704      16882.50   178.5022   720.9559   240.1809  1402.220
#> 1983  23466      24189.41   723.4068  1316.0094  1241.9696  3242.713
#> 1984  27067      28865.12  1798.1199  2025.8318  2707.4013  5597.883
#> 1985  26180      29039.42  2859.4217  2313.8279  3920.0842  7187.412
#> 1986  15852      19549.86  3697.8558  2639.0082  5115.3332  8929.081
#> 1987  12314      17889.32  5575.3195  3079.8227  7318.1120 11358.533
#> 1988  13112      24066.85 10954.8520  4958.5420 13945.5250 19855.447
#> 1989   5395      16034.10 10639.1047  5859.0490 14071.7982 21120.808
#> 1990   2063      18913.07 16850.0704 13512.5823 23830.4142 40609.331
#> 
#> $Totals
#>                    Totals
#> Latest:         160987.00
#> Mean Ultimate:  214263.65
#> Mean IBNR:       53276.65
#> SD IBNR:         18348.11
#> Total IBNR 75%:  62991.19
#> Total IBNR 95%:  86974.74
#> 
mean(B)
#> $ByOrigin
#>       Mean IBNR
#> 1981     0.0000
#> 1982   178.5022
#> 1983   723.4068
#> 1984  1798.1199
#> 1985  2859.4217
#> 1986  3697.8558
#> 1987  5575.3195
#> 1988 10954.8520
#> 1989 10639.1047
#> 1990 16850.0704
#> 
#> $Totals
#>               Total
#> Mean IBNR: 53276.65
#> 
quantile(B, c(0.75,0.95,0.99, 0.995))
#> $ByOrigin
#>        IBNR 75%  IBNR 95%  IBNR 99% IBNR 99.5%
#> 1981     0.0000     0.000     0.000      0.000
#> 1982   240.1809  1402.220  3147.000   3944.181
#> 1983  1241.9696  3242.713  5633.901   6072.495
#> 1984  2707.4013  5597.883  8895.142   9263.866
#> 1985  3920.0842  7187.412  9904.622  11289.317
#> 1986  5115.3332  8929.081 11597.299  13412.057
#> 1987  7318.1120 11358.533 14487.259  15972.350
#> 1988 13945.5250 19855.447 25129.068  26697.154
#> 1989 14071.7982 21120.808 27950.315  30503.814
#> 1990 23830.4142 40609.331 57290.904  67577.719
#> 
#> $Totals
#>                Totals
#> IBNR 75%:    62991.19
#> IBNR 95%:    86974.74
#> IBNR 99%:   105391.55
#> IBNR 99.5%: 113922.34
#>