Methods for BootChainLadder objects

summary, print, mean, and quantile methods for BootChainLadder objects

Usage

# S3 method for BootChainLadder
summary(object, probs=c(0.75,0.95), ...)

# S3 method for BootChainLadder
print(x, probs=c(0.75,0.95), ...)

# S3 method for BootChainLadder
quantile(x, probs=c(0.75, 0.95), na.rm = FALSE,
              names = TRUE, type = 7,...)

# S3 method for BootChainLadder
mean(x, ...)

# S3 method for BootChainLadder
residuals(object, ...)

Arguments

x, object

output from BootChainLadder

probs

numeric vector of probabilities with values in [0,1], see quantile for more help

na.rm

logical; if true, any NA and NaN's are removed from 'x' before the quantiles are computed, see quantile for more help

names

logical; if true, the result has a names attribute. Set to FALSE for speedup with many 'probs', see quantile for more help

type

an integer between 1 and 9 selecting one of the nine quantile algorithms detailed below to be used, see quantile

...

further arguments passed to or from other methods

Details

print.BootChainLadder calls summary.BootChainLadder and prints a formatted version of the summary. residuals.BootChainLadder gives the residual triangle of the expected chain-ladder minus the actual triangle back.

Value

summary.BootChainLadder, mean.BootChainLadder, and

quantile.BootChainLadder, give a list with two elements back:

ByOrigin

data frame with summary/mean/quantile statistics by origin period

Totals

data frame with total summary/mean/quantile statistics for all origin period

Author

Markus Gesmann

See also

Examples

B <- BootChainLadder(RAA, R=999, process.distr="gamma")
B
#> BootChainLadder(Triangle = RAA, R = 999, process.distr = "gamma")
#> 
#>      Latest Mean Ultimate Mean IBNR IBNR.S.E IBNR 75% IBNR 95%
#> 1981 18,834        18,834         0        0        0        0
#> 1982 16,704        16,861       157      719      190    1,357
#> 1983 23,466        24,100       634    1,178    1,107    2,785
#> 1984 27,067        28,906     1,839    1,973    2,840    5,385
#> 1985 26,180        29,114     2,934    2,404    4,022    7,665
#> 1986 15,852        19,523     3,671    2,376    4,918    8,143
#> 1987 12,314        17,706     5,392    3,017    7,112   10,805
#> 1988 13,112        23,985    10,873    4,863   13,737   19,865
#> 1989  5,395        16,731    11,336    6,181   15,072   22,949
#> 1990  2,063        19,490    17,427   13,778   24,897   42,585
#> 
#>                  Totals
#> Latest:         160,987
#> Mean Ultimate:  215,249
#> Mean IBNR:       54,262
#> IBNR.S.E         18,789
#> Total IBNR 75%:  66,195
#> Total IBNR 95%:  88,706
summary(B)
#> $ByOrigin
#>      Latest Mean Ultimate  Mean IBNR    SD IBNR   IBNR 75%  IBNR 95%
#> 1981  18834      18834.00     0.0000     0.0000     0.0000     0.000
#> 1982  16704      16860.97   156.9748   718.8066   189.9167  1357.210
#> 1983  23466      24099.73   633.7314  1178.0252  1106.5649  2785.312
#> 1984  27067      28905.84  1838.8448  1972.6664  2839.9675  5384.743
#> 1985  26180      29113.51  2933.5141  2404.2124  4022.4244  7665.305
#> 1986  15852      19522.84  3670.8446  2376.4408  4918.2899  8142.807
#> 1987  12314      17706.23  5392.2292  3016.9843  7111.5063 10805.210
#> 1988  13112      23984.90 10872.9045  4862.7219 13737.1460 19865.118
#> 1989   5395      16731.01 11336.0123  6180.8178 15072.1285 22948.697
#> 1990   2063      19490.31 17427.3074 13778.1575 24896.7178 42585.485
#> 
#> $Totals
#>                    Totals
#> Latest:         160987.00
#> Mean Ultimate:  215249.36
#> Mean IBNR:       54262.36
#> SD IBNR:         18789.15
#> Total IBNR 75%:  66194.86
#> Total IBNR 95%:  88705.94
#> 
mean(B)
#> $ByOrigin
#>       Mean IBNR
#> 1981     0.0000
#> 1982   156.9748
#> 1983   633.7314
#> 1984  1838.8448
#> 1985  2933.5141
#> 1986  3670.8446
#> 1987  5392.2292
#> 1988 10872.9045
#> 1989 11336.0123
#> 1990 17427.3074
#> 
#> $Totals
#>               Total
#> Mean IBNR: 54262.36
#> 
quantile(B, c(0.75,0.95,0.99, 0.995))
#> $ByOrigin
#>        IBNR 75%  IBNR 95%  IBNR 99% IBNR 99.5%
#> 1981     0.0000     0.000     0.000      0.000
#> 1982   189.9167  1357.210  3098.383   3319.884
#> 1983  1106.5649  2785.312  4804.167   5106.503
#> 1984  2839.9675  5384.743  8122.502   8605.408
#> 1985  4022.4244  7665.305 10484.837  11055.188
#> 1986  4918.2899  8142.807 10518.579  11698.655
#> 1987  7111.5063 10805.210 14348.442  16351.834
#> 1988 13737.1460 19865.118 24131.313  25469.548
#> 1989 15072.1285 22948.697 28804.989  30091.706
#> 1990 24896.7178 42585.485 57267.973  63544.508
#> 
#> $Totals
#>                Totals
#> IBNR 75%:    66194.86
#> IBNR 95%:    88705.94
#> IBNR 99%:   103393.09
#> IBNR 99.5%: 107096.76
#>